@@ -1,18 +1,21 @@
 """Supports reading cyclic structural result files from ANSYS"""
 from functools import wraps
 
-from vtk import vtkMatrix4x4, vtkTransform, vtkAppendFilter
 import numpy as np
 import pyvista as pv
+from pyvista._vtk import vtkAppendFilter, vtkMatrix4x4, vtkTransform
 
-from ansys.mapdl.reader.common import (STRESS_TYPES, STRAIN_TYPES,
-                                       PRINCIPAL_STRESS_TYPES,
-                                       THERMAL_STRAIN_TYPES)
-from ansys.mapdl.reader.rst import Result, check_comp
 from ansys.mapdl.reader import _binary_reader
-from ansys.mapdl.reader.common import axis_rotation
+from ansys.mapdl.reader.common import (
+    PRINCIPAL_STRESS_TYPES,
+    STRAIN_TYPES,
+    STRESS_TYPES,
+    THERMAL_STRAIN_TYPES,
+    axis_rotation,
+)
+from ansys.mapdl.reader.rst import Result, check_comp
 
-np.seterr(divide='ignore', invalid='ignore')
+np.seterr(divide="ignore", invalid="ignore")
 
 
 class CyclicResult(Result):
@@ -24,7 +27,7 @@ class CyclicResult(Result):
 
         # sanity check
         if not self._is_cyclic:
-            raise TypeError('Result is not a cyclic model')
+            raise TypeError("Result is not a cyclic model")
 
         self._rotor_cache = None
         self._has_duplicate_sector = None
@@ -54,8 +57,8 @@ class CyclicResult(Result):
         """
         scalars = np.empty((self.n_sector, self._mas_grid.n_points))
         scalars[:] = np.arange(self.n_sector).reshape(-1, 1)
-        kwargs.setdefault('show_edges', True)
-        kwargs.setdefault('n_colors', self.n_sector)
+        kwargs.setdefault("show_edges", True)
+        kwargs.setdefault("n_colors", self.n_sector)
         return self._plot_cyclic_point_scalars(scalars, None, add_text=False, **kwargs)
 
     def plot(self, **kwargs):
@@ -81,23 +84,23 @@ class CyclicResult(Result):
 
         >>> rst.plot(screenshot='rotor.png')
         """
-        kwargs.setdefault('color', 'w')
-        kwargs.setdefault('show_edges', True)
+        kwargs.setdefault("color", "w")
+        kwargs.setdefault("show_edges", True)
         return self._plot_cyclic_point_scalars(None, None, add_text=False, **kwargs)
 
     def _add_cyclic_properties(self):
         """Add cyclic properties"""
 
         # idenfity the sector based on number of elements in master sector
-        cs_els = self._resultheader['csEls']
-        mask = self.quadgrid.cell_arrays['ansys_elem_num'] <= cs_els
+        cs_els = self._resultheader["csEls"]
+        mask = self.quadgrid.cell_data["ansys_elem_num"] <= cs_els
 
         self.master_cell_mask = mask
         self._mas_grid = self.grid.extract_cells(mask)
 
         # NOTE: number of nodes in sector may not match number of
         # nodes in geometry
-        node_mask = self._neqv[self._sidx] <= self._resultheader['csNds']
+        node_mask = self._neqv[self._sidx] <= self._resultheader["csNds"]
         self._mas_ind = np.nonzero(node_mask)[0]
         self._dup_ind = np.nonzero(~node_mask)[0]
         self._has_duplicate_sector = np.any(self._dup_ind)
@@ -107,6 +110,11 @@ class CyclicResult(Result):
         mask_b = np.isclose(self.time_values, np.roll(self.time_values, -1))
         self._is_repeated_mode = np.logical_or(mask_a, mask_b)
 
+        # should not have repeated modes at harmonic index N/2
+        self._is_repeated_mode[
+            self._resultheader["hindex"] == self.n_sector // 2
+        ] = False
+
         # edge case for single pair of repeated modes
         if self._is_repeated_mode.size == 2 and self._is_repeated_mode.all():
             self._repeated_index = np.array([1, 0])
@@ -115,15 +123,14 @@ class CyclicResult(Result):
             self._is_repeated_mode = np.array([False])
             return
 
-        self._repeated_index = np.empty(self._is_repeated_mode.size, np.int)
+        self._repeated_index = np.empty(self._is_repeated_mode.size, np.int_)
         self._repeated_index[:] = -1
         if np.any(self._is_repeated_mode):
             self._repeated_index[mask_a] = np.nonzero(mask_b)[0]
             self._repeated_index[mask_b] = np.nonzero(mask_a)[0]
 
     def nodal_solution(self, rnum, phase=0, full_rotor=False, as_complex=False):
-        """Returns the DOF solution for each node in the global
-        cartesian coordinate system.
+        """Return the DOF solution for each node in the global cartesian coordinate system.
 
         Parameters
         ----------
@@ -175,8 +182,9 @@ class CyclicResult(Result):
         """
 
         func = super().nodal_solution
-        return self._get_full_result(rnum, func, phase, full_rotor, as_complex,
-                                     tensor=False)
+        return self._get_full_result(
+            rnum, func, phase, full_rotor, as_complex, tensor=False
+        )
 
     @wraps(nodal_solution)
     def nodal_displacement(self, *args, **kwargs):
@@ -185,7 +193,7 @@ class CyclicResult(Result):
 
     def _expand_cyclic_static(self, result, tensor=False, stress=True):
         """Expand cyclic static result for a full rotor"""
-        cs_cord = self._resultheader['csCord']
+        cs_cord = self._resultheader["csCord"]
         if cs_cord > 1:
             matrix = self.cs_4x4(cs_cord, as_vtk_matrix=True)
             i_matrix = self.cs_4x4(cs_cord, as_vtk_matrix=True)
@@ -203,7 +211,7 @@ class CyclicResult(Result):
             # transform to standard position, rotate about Z axis,
             # transform back
             transform = vtkTransform()
-            transform.RotateZ(rang*i)
+            transform.RotateZ(rang * i)
             transform.Update()
             rot_matrix = transform.GetMatrix()
 
@@ -223,19 +231,20 @@ class CyclicResult(Result):
 
         return full_result
 
-    def _expand_cyclic_modal(self, result, result_r, hindex, phase, as_complex,
-                             full_rotor):
-        """ Combines repeated results from ANSYS """
+    def _expand_cyclic_modal(
+        self, result, result_r, hindex, phase, as_complex, full_rotor
+    ):
+        """Combines repeated results from ANSYS"""
         if as_complex or full_rotor:
             # Matches ansys direction
             # if self._positive_cyclic_dir:
-            result_combined = result + result_r*1j
+            result_combined = result + result_r * 1j
             # else:
-                # result_combined = result - result_r*1j
+            # result_combined = result - result_r*1j
             if phase:
-                result_combined *= 1*np.cos(phase) - 1j*np.sin(phase)
+                result_combined *= 1 * np.cos(phase) - 1j * np.sin(phase)
         else:  # convert to real
-            result_combined = result*np.cos(phase) - result_r*np.sin(phase)
+            result_combined = result * np.cos(phase) - result_r * np.sin(phase)
 
         # just return single sector
         if not full_rotor:
@@ -243,35 +252,35 @@ class CyclicResult(Result):
 
         # Generate full rotor solution
         result_expanded = []
-        angles = np.linspace(0, 2*np.pi, self.n_sector + 1)[:-1] + phase
+        angles = np.linspace(0, 2 * np.pi, self.n_sector + 1)[:-1] + phase
         for angle in angles:
             # need to rotate solution and rotate direction
-            result_expanded.append(axis_rotation(result_combined,
-                                                 angle, deg=False,
-                                                 axis='z'))
+            result_expanded.append(
+                axis_rotation(result_combined, angle, deg=False, axis="z")
+            )
 
         result_expanded = np.asarray(result_expanded, dtype=np.complex128)
 
         # ANSYS scales the result
-        if hindex == 0 or hindex == self.n_sector/2:
+        if hindex == 0 or hindex == self.n_sector / 2:
             result_expanded /= self.n_sector**0.5
         else:
-            result_expanded /= (self.n_sector/2)**0.5
+            result_expanded /= (self.n_sector / 2) ** 0.5
 
         # adjust phase of the full result based on the harmonic index
         f_arr = np.zeros(self.n_sector)
         f_arr[hindex] = 1
-        jang = np.fft.ifft(f_arr)[:self.n_sector]*self.n_sector
+        jang = np.fft.ifft(f_arr)[: self.n_sector] * self.n_sector
         cjang = jang * (np.cos(phase) - np.sin(phase) * 1j)  # 14-233
 
         result_expanded *= cjang.reshape(-1, 1, 1)
         if as_complex:
             return result_expanded
-        else:
-            return np.real(result_expanded)
+        return np.real(result_expanded)
 
-    def _expand_cyclic_modal_tensor(self, result, result_r, hindex,
-                                    phase, as_complex, full_rotor, stress=True):
+    def _expand_cyclic_modal_tensor(
+        self, result, result_r, hindex, phase, as_complex, full_rotor, stress=True
+    ):
         """Combines repeated results from ANSYS and optionally
         duplicates/rotates it around the axis of rotation"""
 
@@ -279,15 +288,15 @@ class CyclicResult(Result):
 
         if as_complex or full_rotor:
             # if self._positive_cyclic_dir:
-            result_combined = result + result_r*1j
+            result_combined = result + result_r * 1j
             # else:
-                # result_combined = result - result_r*1j
+            # result_combined = result - result_r*1j
 
             if phase:
-                result_combined *= 1*np.cos(phase) - 1j*np.sin(phase)
+                result_combined *= 1 * np.cos(phase) - 1j * np.sin(phase)
 
         else:  # convert to real
-            result_combined = result*np.cos(phase) - result_r*np.sin(phase)
+            result_combined = result * np.cos(phase) - result_r * np.sin(phase)
 
         # just return single sector
         if not full_rotor:
@@ -295,27 +304,27 @@ class CyclicResult(Result):
 
         # Generate full rotor solution
         shp = (self.n_sector, result.shape[0], result.shape[1])
-        result_expanded = np.empty(shp,np.complex128)
+        result_expanded = np.empty(shp, np.complex128)
         result_expanded[:] = result_combined
 
         # convert hindex to nodal content
         f_arr = np.zeros(self.n_sector)
         f_arr[hindex] = self.n_sector
 
-        jang = np.fft.ifft(f_arr)[:self.n_sector]
+        jang = np.fft.ifft(f_arr)[: self.n_sector]
 
         # must be adjusted of to expand
-        if hindex == 0 or hindex == self.n_sector/2:
-            jang *= self.n_sector**(-1/2)
+        if hindex == 0 or hindex == self.n_sector / 2:
+            jang *= self.n_sector ** (-1 / 2)
         else:
-            jang *= 2*(2*self.n_sector)**(-1/2)
+            jang *= 2 * (2 * self.n_sector) ** (-1 / 2)
 
         cjang = jang * (np.cos(phase) - np.sin(phase) * 1j)
 
         full_result = np.empty(shp)
-        full_result[:] = np.real(result_expanded*cjang.reshape(-1, 1, 1))
+        full_result[:] = np.real(result_expanded * cjang.reshape(-1, 1, 1))
 
-        cs_cord = self._resultheader['csCord']
+        cs_cord = self._resultheader["csCord"]
         if cs_cord > 1:
             matrix = self.cs_4x4(cs_cord, as_vtk_matrix=True)
             i_matrix = self.cs_4x4(cs_cord, as_vtk_matrix=True)
@@ -329,7 +338,7 @@ class CyclicResult(Result):
             # transform to standard position, rotate about Z axis,
             # transform back
             transform = vtkTransform()
-            transform.RotateZ(rang*i)
+            transform.RotateZ(rang * i)
             transform.Update()
             rot_matrix = transform.GetMatrix()
 
@@ -369,28 +378,31 @@ class CyclicResult(Result):
             Cumulative index number.  Zero based indexing.
 
         """
-        hindex_table = self._resultheader['hindex']
+        hindex_table = self._resultheader["hindex"]
         if not np.any(abs(hindex) == np.abs(hindex_table)):
-            raise Exception('Invalid harmonic index.\n' +
-                            'Available indices: %s' % np.unique(hindex_table))
+            raise Exception(
+                "Invalid harmonic index.\n"
+                + "Available indices: %s" % np.unique(hindex_table)
+            )
 
-        mask = np.logical_and(hindex == hindex_table,
-                              mode == self.mode_table)
+        mask = np.logical_and(hindex == hindex_table, mode == self.mode_table)
 
         if not mask.any():
             mode_mask = abs(hindex) == np.abs(hindex_table)
             avail_modes = np.unique(self.mode_table[mode_mask])
-            raise Exception('Invalid mode for harmonic index %d\n' % hindex +
-                            'Available modes: %s' % avail_modes)
+            raise Exception(
+                "Invalid mode for harmonic index %d\n" % hindex
+                + "Available modes: %s" % avail_modes
+            )
 
         index = mask.nonzero()[0]
-        assert index.size == 1, 'Multiple cumulative index matches'
+        assert index.size == 1, "Multiple cumulative index matches"
         return index[0]
 
     @property
     def mode_table(self):
         """Unique modes for cyclic results"""
-        hindex_table = self._resultheader['hindex']
+        hindex_table = self._resultheader["hindex"]
         diff = np.diff(np.abs(hindex_table))
         freqs = self.time_values
         mode_table = [0]
@@ -420,7 +432,7 @@ class CyclicResult(Result):
         array([ 0,  0,  0,  0,  0,  0, -1,  1, -1,  1,  1, -1,
                -2,  2, -2,  2, -2,  2,  3,  3,  3,  3,  3,  3], dtype=int32)
         """
-        return self._resultheader['hindex']
+        return self._resultheader["hindex"]
 
     def nodal_stress(self, rnum, phase=0, as_complex=False, full_rotor=False):
         """Retrieves the component stresses for each node in the
@@ -473,11 +485,13 @@ class CyclicResult(Result):
 
         """
         func = super().nodal_stress
-        return self._get_full_result(rnum, func, phase, full_rotor, as_complex,
-                                     tensor=True, stress=True)
+        return self._get_full_result(
+            rnum, func, phase, full_rotor, as_complex, tensor=True, stress=True
+        )
 
-    def _get_full_result(self, rnum, func, phase, full_rotor, as_complex,
-                         tensor=True, stress=False):
+    def _get_full_result(
+        self, rnum, func, phase, full_rotor, as_complex, tensor=True, stress=False
+    ):
         """Return the full rotor result or the complex result for a cyclic model.
 
         rnum : int or list
@@ -501,41 +515,47 @@ class CyclicResult(Result):
 
         # full result may or may not contain the duplicate sector
         if self._has_duplicate_sector:
-            result = full_result[self._mas_ind]
-            nnum = nnum[self._mas_ind]
+            if nnum.size < self._mas_ind.size:
+                # node numbers of the master sector
+                nnum_mas = self._neqv[self._mas_ind]
+                mask = np.in1d(nnum, nnum_mas)
+                nnum = nnum[mask]
+                result = full_result[mask]
+            else:
+                result = full_result[self._mas_ind]
+                nnum = nnum[self._mas_ind]
         else:
             result = full_result
 
-        if self._resultheader['kan'] == 0:  # static result
+        if self._resultheader["kan"] == 0:  # static result
             if full_rotor:
-                expanded_result = self._expand_cyclic_static(result,
-                                                             tensor=tensor,
-                                                             stress=stress)
+                expanded_result = self._expand_cyclic_static(
+                    result, tensor=tensor, stress=stress
+                )
             else:
                 return nnum, result
-        elif self._resultheader['kan'] == 2:  # modal analysis
+        elif self._resultheader["kan"] == 2:  # modal analysis
             # combine modal solution results
-            hindex_table = self._resultheader['hindex']
+            hindex_table = self._resultheader["hindex"]
             hindex = hindex_table[rnum]  # move this to expand_modal_tensor
             result_r = self._get_complex_result(func, rnum, result)
             if tensor:
-                expanded_result = self._expand_cyclic_modal_tensor(result,
-                                                                   result_r,
-                                                                   hindex,
-                                                                   phase,
-                                                                   as_complex,
-                                                                   full_rotor,
-                                                                   stress=stress)
+                expanded_result = self._expand_cyclic_modal_tensor(
+                    result,
+                    result_r,
+                    hindex,
+                    phase,
+                    as_complex,
+                    full_rotor,
+                    stress=stress,
+                )
             else:
                 assert result.shape == result_r.shape
-                expanded_result = self._expand_cyclic_modal(result,
-                                                            result_r,
-                                                            hindex,
-                                                            phase,
-                                                            as_complex,
-                                                            full_rotor)
+                expanded_result = self._expand_cyclic_modal(
+                    result, result_r, hindex, phase, as_complex, full_rotor
+                )
         else:
-            raise RuntimeError('Unsupported analysis type')
+            raise RuntimeError("Unsupported analysis type")
 
         return nnum, expanded_result
 
@@ -575,9 +595,10 @@ class CyclicResult(Result):
                 result_r = result_r[self._mas_ind]
         elif has_dup_result:  # use the duplicate sector
             result_r = full_result[self._dup_ind]
-        else:  # otherwise, a standing wave (no complex component)            
-            result_r = np.zeros((self._mas_ind.size, full_result.shape[1]),
-                                dtype=full_result.dtype)
+        else:  # otherwise, a standing wave (no complex component)
+            result_r = np.zeros(
+                (self._mas_ind.size, full_result.shape[1]), dtype=full_result.dtype
+            )
 
         return result_r
 
@@ -627,22 +648,26 @@ class CyclicResult(Result):
 
         """
         func = super().nodal_thermal_strain
-        return self._get_full_result(rnum, func, phase, full_rotor, as_complex,
-                                     tensor=True, stress=False)
+        return self._get_full_result(
+            rnum, func, phase, full_rotor, as_complex, tensor=True, stress=False
+        )
 
-    def plot_nodal_thermal_strain(self, rnum,
-                                  comp=None,
-                                  phase=0,
-                                  full_rotor=True,
-                                  show_displacement=False,
-                                  displacement_factor=1,
-                                  node_components=None,
-                                  element_components=None,
-                                  sel_type_all=True,
-                                  add_text=True,
-                                  overlay_wireframe=False,
-                                  treat_nan_as_zero=True,
-                                  **kwargs):
+    def plot_nodal_thermal_strain(
+        self,
+        rnum,
+        comp=None,
+        phase=0,
+        full_rotor=True,
+        show_displacement=False,
+        displacement_factor=1,
+        node_components=None,
+        element_components=None,
+        sel_type_all=True,
+        add_text=True,
+        overlay_wireframe=False,
+        treat_nan_as_zero=False,
+        **kwargs,
+    ):
         """Plot nodal thermal strain.
 
         Parameters
@@ -706,32 +731,30 @@ class CyclicResult(Result):
 
         """
         if not full_rotor:
-            return super().plot_nodal_thermal_strain(rnum,
-                                                     treat_nan_as_zero=treat_nan_as_zero,
-                                                     **kwargs)
+            return super().plot_nodal_thermal_strain(
+                rnum, treat_nan_as_zero=treat_nan_as_zero, **kwargs
+            )
 
         idx = check_comp(THERMAL_STRAIN_TYPES, comp)
         _, strain = self.nodal_thermal_strain(rnum, phase, False, True)
         scalars = strain[:, :, idx]
 
-        kwargs.setdefault('scalar_bar_args', {'title': f'{comp} Nodal Thermal Strain'})
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['show_displacement'] = show_displacement
-        kwargs['displacement_factor'] = displacement_factor
-        kwargs['overlay_wireframe'] = overlay_wireframe
-        kwargs['add_text'] = add_text
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['sel_type_all'] = sel_type_all
-        kwargs['phase'] = phase
-        return self._plot_cyclic_point_scalars(scalars,
-                                               rnum,
-                                               treat_nan_as_zero=treat_nan_as_zero,
-                                               **kwargs)
+        kwargs.setdefault("scalar_bar_args", {"title": f"{comp} Nodal Thermal Strain"})
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["show_displacement"] = show_displacement
+        kwargs["displacement_factor"] = displacement_factor
+        kwargs["overlay_wireframe"] = overlay_wireframe
+        kwargs["add_text"] = add_text
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["sel_type_all"] = sel_type_all
+        kwargs["phase"] = phase
+        return self._plot_cyclic_point_scalars(
+            scalars, rnum, treat_nan_as_zero=treat_nan_as_zero, **kwargs
+        )
 
-    def nodal_elastic_strain(self, rnum, phase=0, as_complex=False,
-                             full_rotor=False):
+    def nodal_elastic_strain(self, rnum, phase=0, as_complex=False, full_rotor=False):
         """Nodal component elastic strains.  This record contains
         strains in the order X, Y, Z, XY, YZ, XZ, EQV.
 
@@ -779,21 +802,26 @@ class CyclicResult(Result):
 
         """
         func = super().nodal_elastic_strain
-        return self._get_full_result(rnum, func, phase, full_rotor, as_complex,
-                                     tensor=True, stress=False)
+        return self._get_full_result(
+            rnum, func, phase, full_rotor, as_complex, tensor=True, stress=False
+        )
 
-    def plot_nodal_elastic_strain(self, rnum,
-                                  comp=None, phase=0,
-                                  full_rotor=True,
-                                  show_displacement=False,
-                                  displacement_factor=1,
-                                  node_components=None,
-                                  element_components=None,
-                                  sel_type_all=True,
-                                  add_text=True,
-                                  overlay_wireframe=False,
-                                  treat_nan_as_zero=True,
-                                  **kwargs):
+    def plot_nodal_elastic_strain(
+        self,
+        rnum,
+        comp=None,
+        phase=0,
+        full_rotor=True,
+        show_displacement=False,
+        displacement_factor=1,
+        node_components=None,
+        element_components=None,
+        sel_type_all=True,
+        add_text=True,
+        overlay_wireframe=False,
+        treat_nan_as_zero=False,
+        **kwargs,
+    ):
         """Plot nodal elastic strain.
 
         Parameters
@@ -858,24 +886,22 @@ class CyclicResult(Result):
         _, strain = self.nodal_elastic_strain(rnum, phase, False, full_rotor)
         scalars = strain[:, :, idx]
 
-        kwargs.setdefault('scalar_bar_args', {'title': f'{comp} Nodal Elastic Strain'})
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['show_displacement'] = show_displacement
-        kwargs['displacement_factor'] = displacement_factor
-        kwargs['overlay_wireframe'] = overlay_wireframe
-        kwargs['add_text'] = add_text
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['sel_type_all'] = sel_type_all
-        kwargs['phase'] = phase
-        return self._plot_cyclic_point_scalars(scalars,
-                                               rnum,
-                                               treat_nan_as_zero=treat_nan_as_zero,
-                                               **kwargs)
+        kwargs.setdefault("scalar_bar_args", {"title": f"{comp} Nodal Elastic Strain"})
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["show_displacement"] = show_displacement
+        kwargs["displacement_factor"] = displacement_factor
+        kwargs["overlay_wireframe"] = overlay_wireframe
+        kwargs["add_text"] = add_text
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["sel_type_all"] = sel_type_all
+        kwargs["phase"] = phase
+        return self._plot_cyclic_point_scalars(
+            scalars, rnum, treat_nan_as_zero=treat_nan_as_zero, **kwargs
+        )
 
-    def nodal_plastic_strain(self, rnum, phase=0, as_complex=False,
-                             full_rotor=False):
+    def nodal_plastic_strain(self, rnum, phase=0, as_complex=False, full_rotor=False):
         """Nodal component plastic strains.  This record contains
         strains in the order X, Y, Z, XY, YZ, XZ, EQV.
 
@@ -923,21 +949,26 @@ class CyclicResult(Result):
 
         """
         func = super().nodal_plastic_strain
-        return self._get_full_result(rnum, func, phase, full_rotor, as_complex,
-                                     tensor=True, stress=False)
+        return self._get_full_result(
+            rnum, func, phase, full_rotor, as_complex, tensor=True, stress=False
+        )
 
-    def plot_nodal_plastic_strain(self, rnum,
-                                  comp=None, phase=0,
-                                  full_rotor=True,
-                                  show_displacement=False,
-                                  displacement_factor=1,
-                                  node_components=None,
-                                  element_components=None,
-                                  sel_type_all=True,
-                                  add_text=True,
-                                  overlay_wireframe=False,
-                                  treat_nan_as_zero=True,
-                                  **kwargs):
+    def plot_nodal_plastic_strain(
+        self,
+        rnum,
+        comp=None,
+        phase=0,
+        full_rotor=True,
+        show_displacement=False,
+        displacement_factor=1,
+        node_components=None,
+        element_components=None,
+        sel_type_all=True,
+        add_text=True,
+        overlay_wireframe=False,
+        treat_nan_as_zero=False,
+        **kwargs,
+    ):
         """Plot nodal plastic strain.
 
         Parameters
@@ -1002,24 +1033,22 @@ class CyclicResult(Result):
         _, strain = self.nodal_plastic_strain(rnum, phase, False, full_rotor)
         scalars = strain[:, :, idx]
 
-        kwargs.setdefault('scalar_bar_args', {'title': f'{comp} Nodal Plastic Strain'})
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['show_displacement'] = show_displacement
-        kwargs['displacement_factor'] = displacement_factor
-        kwargs['overlay_wireframe'] = overlay_wireframe
-        kwargs['add_text'] = add_text
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['sel_type_all'] = sel_type_all
-        kwargs['phase'] = phase
-        return self._plot_cyclic_point_scalars(scalars,
-                                               rnum,
-                                               treat_nan_as_zero=treat_nan_as_zero,
-                                               **kwargs)
+        kwargs.setdefault("scalar_bar_args", {"title": f"{comp} Nodal Plastic Strain"})
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["show_displacement"] = show_displacement
+        kwargs["displacement_factor"] = displacement_factor
+        kwargs["overlay_wireframe"] = overlay_wireframe
+        kwargs["add_text"] = add_text
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["sel_type_all"] = sel_type_all
+        kwargs["phase"] = phase
+        return self._plot_cyclic_point_scalars(
+            scalars, rnum, treat_nan_as_zero=treat_nan_as_zero, **kwargs
+        )
 
-    def principal_nodal_stress(self, rnum, phase=0, as_complex=False,
-                               full_rotor=False):
+    def principal_nodal_stress(self, rnum, phase=0, as_complex=False, full_rotor=False):
         """Computes the principal component stresses for each node in
         the solution.
 
@@ -1058,7 +1087,7 @@ class CyclicResult(Result):
         equivalent stress.
         """
         if as_complex and full_rotor:
-            raise ValueError('complex and full_rotor cannot both be True')
+            raise ValueError("`complex` and `full_rotor` cannot both be True")
 
         # get component stress
         nnum, stress = self.nodal_stress(rnum, phase, as_complex, full_rotor)
@@ -1068,14 +1097,19 @@ class CyclicResult(Result):
             stress_r = np.imag(stress)
             stress = np.real(stress)
 
+            if not stress.flags["C_CONTIGUOUS"]:
+                stress = np.ascontiguousarray(stress)
+            if not stress_r.flags["C_CONTIGUOUS"]:
+                stress_r = np.ascontiguousarray(stress_r)
+
             pstress, isnan = _binary_reader.compute_principal_stress(stress)
             pstress[isnan] = np.nan
             pstress_r, isnan = _binary_reader.compute_principal_stress(stress_r)
             pstress_r[isnan] = np.nan
 
-            return nnum, pstress + 1j*pstress_r
+            return nnum, pstress + 1j * pstress_r
 
-        elif full_rotor:
+        if full_rotor:
             # compute principle stress for each sector
             pstress = np.empty((self.n_sector, stress.shape[1], 5), np.float64)
             for i in range(stress.shape[0]):
@@ -1083,23 +1117,26 @@ class CyclicResult(Result):
                 pstress[i, isnan] = np.nan
             return nnum, pstress
 
-        else:
-            pstress, isnan = _binary_reader.compute_principal_stress(stress)
-            pstress[isnan] = np.nan
-            return nnum, pstress
+        pstress, isnan = _binary_reader.compute_principal_stress(stress)
+        pstress[isnan] = np.nan
+        return nnum, pstress
 
-    def plot_nodal_solution(self, rnum, comp='norm',
-                            phase=0,
-                            full_rotor=True,
-                            show_displacement=False,
-                            displacement_factor=1.0,
-                            node_components=None,
-                            element_components=None,
-                            overlay_wireframe=False,
-                            add_text=True,
-                            sel_type_all=True,
-                            treat_nan_as_zero=True,
-                            **kwargs):
+    def plot_nodal_solution(
+        self,
+        rnum,
+        comp="norm",
+        phase=0,
+        full_rotor=True,
+        show_displacement=False,
+        displacement_factor=1.0,
+        node_components=None,
+        element_components=None,
+        overlay_wireframe=False,
+        add_text=True,
+        sel_type_all=True,
+        treat_nan_as_zero=False,
+        **kwargs,
+    ):
         """Plot the nodal solution (generally displacement).
 
         Parameters
@@ -1110,7 +1147,7 @@ class CyclicResult(Result):
 
         comp : str, optional
             Display component to display.  Options are 'x', 'y', 'z',
-            and 'norm', corresponding to the x directin, y direction,
+            and 'norm', corresponding to the x direction, y direction,
             z direction, and the normalized direction:
             ``(x**2 + y**2 + z**2)**0.5``
 
@@ -1151,67 +1188,70 @@ class CyclicResult(Result):
         >>> result.plot_nodal_solution(0)
 
         """
-
         # Load result from file
         if not full_rotor:
-            return super().plot_nodal_solution(rnum,
-                                               comp,
-                                               show_displacement=show_displacement,
-                                               displacement_factor=displacement_factor,
-                                               node_components=node_components,
-                                               element_components=element_components,
-                                               sel_type_all=sel_type_all,
-                                               treat_nan_as_zero=treat_nan_as_zero,
-                                               **kwargs)
+            return super().plot_nodal_solution(
+                rnum,
+                comp,
+                show_displacement=show_displacement,
+                displacement_factor=displacement_factor,
+                node_components=node_components,
+                element_components=element_components,
+                sel_type_all=sel_type_all,
+                treat_nan_as_zero=treat_nan_as_zero,
+                **kwargs,
+            )
 
         rnum = self.parse_step_substep(rnum)
         _, result = self.nodal_solution(rnum, phase, full_rotor, as_complex=False)
 
         # Process result
-        label = 'Cyclic Rotor\nDisplacement'
-        if comp == 'x':
+        label = "Cyclic Rotor\nDisplacement"
+        if comp == "x":
             scalars = result[:, :, 0]
-            title = 'X {:s}\n'.format(label)
-        elif comp == 'y':
+            title = f"X {label}\n"
+        elif comp == "y":
             scalars = result[:, :, 1]
-            title = 'Y {:s}\n'.format(label)
-        elif comp == 'z':
+            title = f"Y {label}\n"
+        elif comp == "z":
             scalars = result[:, :, 2]
-            title = 'Z {:s}\n'.format(label)
+            title = f"Z {label}\n"
         else:
             # Normalize displacement
-            scalars = (result*result).sum(2)**0.5
-            title = 'Normalized\n%s\n' % label
+            scalars = np.linalg.norm(result, axis=-1)
+            title = f"Normalized\n{label}\n"
 
-        kwargs.setdefault('scalar_bar_args', {'title': title})
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['show_displacement'] = show_displacement
-        kwargs['displacement_factor'] = displacement_factor
-        kwargs['overlay_wireframe'] = overlay_wireframe
-        kwargs['add_text'] = add_text
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['sel_type_all'] = sel_type_all
-        kwargs['phase'] = phase
-        return self._plot_cyclic_point_scalars(scalars,
-                                               rnum,
-                                               treat_nan_as_zero=treat_nan_as_zero,
-                                               **kwargs)
+        kwargs.setdefault("scalar_bar_args", {"title": title})
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["show_displacement"] = show_displacement
+        kwargs["displacement_factor"] = displacement_factor
+        kwargs["overlay_wireframe"] = overlay_wireframe
+        kwargs["add_text"] = add_text
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["sel_type_all"] = sel_type_all
+        kwargs["phase"] = phase
+        return self._plot_cyclic_point_scalars(
+            scalars, rnum, treat_nan_as_zero=treat_nan_as_zero, **kwargs
+        )
 
-    def plot_nodal_stress(self, rnum,
-                          comp=None,
-                          phase=0,
-                          full_rotor=True,
-                          show_displacement=False,
-                          displacement_factor=1,
-                          node_components=None,
-                          element_components=None,
-                          overlay_wireframe=False,
-                          add_text=True,
-                          sel_type_all=True,
-                          treat_nan_as_zero=True,
-                          **kwargs):
+    def plot_nodal_stress(
+        self,
+        rnum,
+        comp=None,
+        phase=0,
+        full_rotor=True,
+        show_displacement=False,
+        displacement_factor=1,
+        node_components=None,
+        element_components=None,
+        overlay_wireframe=False,
+        add_text=True,
+        sel_type_all=True,
+        treat_nan_as_zero=False,
+        **kwargs,
+    ):
         """Plot nodal stress of a given component
 
         Parameters
@@ -1273,49 +1313,55 @@ class CyclicResult(Result):
 
         """
         if not full_rotor:
-            super().plot_nodal_stress(rnum, comp,
-                                      show_displacement,
-                                      displacement_factor,
-                                      node_components,
-                                      element_components,
-                                      sel_type_all,
-                                      treat_nan_as_zero=treat_nan_as_zero,
-                                      **kwargs)
+            super().plot_nodal_stress(
+                rnum,
+                comp,
+                show_displacement,
+                displacement_factor,
+                node_components,
+                element_components,
+                sel_type_all,
+                treat_nan_as_zero=treat_nan_as_zero,
+                **kwargs,
+            )
 
         idx = check_comp(STRESS_TYPES, comp)
         _, stress = self.nodal_stress(rnum, phase, False, full_rotor=True)
         scalars = stress[:, :, idx]
 
-        kwargs.setdefault('scalar_bar_args',
-                          {'title': f'Cyclic Rotor\nNodal Stress\n{comp}\n'})
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['show_displacement'] = show_displacement
-        kwargs['displacement_factor'] = displacement_factor
-        kwargs['overlay_wireframe'] = overlay_wireframe
-        kwargs['add_text'] = add_text
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['sel_type_all'] = sel_type_all
-        kwargs['phase'] = phase
-        return self._plot_cyclic_point_scalars(scalars,
-                                               rnum,
-                                               treat_nan_as_zero=treat_nan_as_zero,
-                                               **kwargs)
+        kwargs.setdefault(
+            "scalar_bar_args", {"title": f"Cyclic Rotor\nNodal Stress\n{comp}\n"}
+        )
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["show_displacement"] = show_displacement
+        kwargs["displacement_factor"] = displacement_factor
+        kwargs["overlay_wireframe"] = overlay_wireframe
+        kwargs["add_text"] = add_text
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["sel_type_all"] = sel_type_all
+        kwargs["phase"] = phase
+        return self._plot_cyclic_point_scalars(
+            scalars, rnum, treat_nan_as_zero=treat_nan_as_zero, **kwargs
+        )
 
-    def plot_principal_nodal_stress(self, rnum,
-                                    comp=None,
-                                    phase=0,
-                                    full_rotor=True,
-                                    show_displacement=False,
-                                    displacement_factor=1,
-                                    node_components=None,
-                                    element_components=None,
-                                    sel_type_all=True,
-                                    add_text=True,
-                                    overlay_wireframe=False,
-                                    treat_nan_as_zero=True,
-                                    **kwargs):
+    def plot_principal_nodal_stress(
+        self,
+        rnum,
+        comp=None,
+        phase=0,
+        full_rotor=True,
+        show_displacement=False,
+        displacement_factor=1,
+        node_components=None,
+        element_components=None,
+        sel_type_all=True,
+        add_text=True,
+        overlay_wireframe=False,
+        treat_nan_as_zero=False,
+        **kwargs,
+    ):
         """Plot the nodal principal stress.
 
         Parameters
@@ -1373,36 +1419,40 @@ class CyclicResult(Result):
 
         """
         if not full_rotor:
-            return super().plot_principal_nodal_stress(rnum, comp,
-                                                       show_displacement,
-                                                       displacement_factor,
-                                                       node_components,
-                                                       element_components,
-                                                       sel_type_all,
-                                                       treat_nan_as_zero=treat_nan_as_zero,
-                                                       **kwargs)
+            return super().plot_principal_nodal_stress(
+                rnum,
+                comp,
+                show_displacement,
+                displacement_factor,
+                node_components,
+                element_components,
+                sel_type_all,
+                treat_nan_as_zero=treat_nan_as_zero,
+                **kwargs,
+            )
 
         # get the correct component of the principal stress for the rotor
         idx = check_comp(PRINCIPAL_STRESS_TYPES, comp)
         _, pstress = self.principal_nodal_stress(rnum, phase, full_rotor=True)
         scalars = pstress[:, :, idx]
 
-        kwargs.setdefault('scalar_bar_args',
-                          {'title': f'Cyclic Rotor\nPrincipal Nodal Stress\n{comp}\n'})
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['show_displacement'] = show_displacement
-        kwargs['displacement_factor'] = displacement_factor
-        kwargs['overlay_wireframe'] = overlay_wireframe
-        kwargs['add_text'] = add_text
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['sel_type_all'] = sel_type_all
-        kwargs['phase'] = phase
-        self._plot_cyclic_point_scalars(scalars,
-                                        rnum,
-                                        treat_nan_as_zero=treat_nan_as_zero,
-                                        **kwargs)
+        kwargs.setdefault(
+            "scalar_bar_args",
+            {"title": f"Cyclic Rotor\nPrincipal Nodal Stress\n{comp}\n"},
+        )
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["show_displacement"] = show_displacement
+        kwargs["displacement_factor"] = displacement_factor
+        kwargs["overlay_wireframe"] = overlay_wireframe
+        kwargs["add_text"] = add_text
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["sel_type_all"] = sel_type_all
+        kwargs["phase"] = phase
+        self._plot_cyclic_point_scalars(
+            scalars, rnum, treat_nan_as_zero=treat_nan_as_zero, **kwargs
+        )
 
     def nodal_temperature(self, rnum, full_rotor=False):
         """Retrieves the temperature for each node in the solution.
@@ -1436,7 +1486,7 @@ class CyclicResult(Result):
         >>> nnum, stress = rst.nodal_temperature(0)
 
         """
-        nnum, temp = super()._nodal_result(rnum, 'EPT')
+        nnum, temp = super()._nodal_result(rnum, "EPT")
         nnum = nnum[self._mas_ind]
         temp = temp[self._mas_ind]
         if not full_rotor:  # return only the master sector
@@ -1445,18 +1495,21 @@ class CyclicResult(Result):
         # otherwise, duplicate and repeat as temperature is constant across sectors
         return nnum, temp.T.repeat(self.n_sector, axis=0)
 
-    def plot_nodal_temperature(self, rnum,
-                               phase=0,
-                               full_rotor=True,
-                               show_displacement=False,
-                               displacement_factor=1.0,
-                               node_components=None,
-                               overlay_wireframe=False,
-                               add_text=True,
-                               element_components=None,
-                               sel_type_all=True,
-                               treat_nan_as_zero=True,
-                               **kwargs):
+    def plot_nodal_temperature(
+        self,
+        rnum,
+        phase=0,
+        full_rotor=True,
+        show_displacement=False,
+        displacement_factor=1.0,
+        node_components=None,
+        overlay_wireframe=False,
+        add_text=True,
+        element_components=None,
+        sel_type_all=True,
+        treat_nan_as_zero=False,
+        **kwargs,
+    ):
         """Plot the nodal temperature.
 
         Parameters
@@ -1504,38 +1557,44 @@ class CyclicResult(Result):
         """
         # Load result from file
         if not full_rotor:
-            return super().plot_nodal_temperature(rnum,
-                                                  show_displacement=show_displacement,
-                                                  displacement_factor=displacement_factor,
-                                                  node_components=node_components,
-                                                  element_components=element_components,
-                                                  sel_type_all=sel_type_all,
-                                                  treat_nan_as_zero=treat_nan_as_zero,
-                                                  **kwargs)
+            return super().plot_nodal_temperature(
+                rnum,
+                show_displacement=show_displacement,
+                displacement_factor=displacement_factor,
+                node_components=node_components,
+                element_components=element_components,
+                sel_type_all=sel_type_all,
+                treat_nan_as_zero=treat_nan_as_zero,
+                **kwargs,
+            )
 
         _, temp = self.nodal_temperature(rnum, True)
 
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['show_displacement'] = show_displacement
-        kwargs['displacement_factor'] = displacement_factor
-        kwargs['overlay_wireframe'] = overlay_wireframe
-        kwargs['add_text'] = add_text
-        kwargs['node_components'] = node_components
-        kwargs['element_components'] = element_components
-        kwargs['sel_type_all'] = sel_type_all
-        kwargs['phase'] = phase
-        return self._plot_cyclic_point_scalars(temp,
-                                               rnum,
-                                               treat_nan_as_zero=treat_nan_as_zero,
-                                               **kwargs)
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["show_displacement"] = show_displacement
+        kwargs["displacement_factor"] = displacement_factor
+        kwargs["overlay_wireframe"] = overlay_wireframe
+        kwargs["add_text"] = add_text
+        kwargs["node_components"] = node_components
+        kwargs["element_components"] = element_components
+        kwargs["sel_type_all"] = sel_type_all
+        kwargs["phase"] = phase
+        return self._plot_cyclic_point_scalars(
+            temp, rnum, treat_nan_as_zero=treat_nan_as_zero, **kwargs
+        )
 
-    def animate_nodal_solution(self, rnum, comp='norm',
-                               displacement_factor=0.1,
-                               nangles=180,
-                               add_text=True, loop=True,
-                               movie_filename=None,
-                               **kwargs):
+    def animate_nodal_solution(
+        self,
+        rnum,
+        comp="norm",
+        displacement_factor=0.1,
+        n_frames=180,
+        add_text=True,
+        loop=True,
+        movie_filename=None,
+        **kwargs,
+    ):
         """Animate nodal solution.  Assumes nodal solution is a
         displacement array from a modal solution.
 
@@ -1551,7 +1610,7 @@ class CyclicResult(Result):
         displacement_factor : float, optional
             Increases or decreases displacement by a factor.
 
-        nangles : int, optional
+        n_frames : int, optional
             Number of "frames" between each full cycle.
 
         show_phase : bool, optional
@@ -1573,26 +1632,32 @@ class CyclicResult(Result):
             See help(pyvista.plot) for additional keyword arguments.
 
         """
+        if "nangles" in kwargs:  # pragma: no cover
+            n_frames = kwargs.pop("nangles")
+            warnings.warn(
+                "The ``nangles`` kwarg is depreciated and ``n_frames`` "
+                "should be used instead."
+            )
+
         rnum = self.parse_step_substep(rnum)  # need cumulative
-        if 'full_rotor' in kwargs:
-            raise NotImplementedError('``full_rotor`` keyword argument not supported')
+        if "full_rotor" in kwargs:
+            raise NotImplementedError("``full_rotor`` keyword argument not supported")
 
         # normalize nodal solution
-        _, complex_disp = self.nodal_solution(rnum, as_complex=True,
-                                              full_rotor=True)
+        _, complex_disp = self.nodal_solution(rnum, as_complex=True, full_rotor=True)
         complex_disp *= displacement_factor
         complex_disp = complex_disp.reshape(-1, 3)
 
-        if comp == 'x':
+        if comp == "x":
             axis = 0
-        elif comp == 'y':
+        elif comp == "y":
             axis = 1
-        elif comp == 'z':
+        elif comp == "z":
             axis = 2
         else:
             axis = None
 
-        result_info = ''
+        result_info = ""
         if add_text:
             result_info = self.text_result_table(rnum)
 
@@ -1601,50 +1666,50 @@ class CyclicResult(Result):
         orig_pt = plot_mesh.points.copy()
 
         # reduce the complex displacement to just the surface points
-        ind = plot_mesh.point_arrays['vtkOriginalPointIds']
+        ind = plot_mesh.point_data["vtkOriginalPointIds"]
         complex_disp = np.take(complex_disp, ind, axis=0)
 
         if axis is not None:
             scalars = complex_disp[:, axis]
         else:
-            scalars = (complex_disp*complex_disp).sum(1)**0.5
+            scalars = (complex_disp * complex_disp).sum(1) ** 0.5
 
         # initialize plotter
-        text_color = kwargs.pop('text_color', None)
-        cpos = kwargs.pop('cpos', None)
-        off_screen = kwargs.pop('off_screen', None)
+        text_color = kwargs.pop("text_color", None)
+        cpos = kwargs.pop("cpos", None)
+        off_screen = kwargs.pop("off_screen", None)
         plotter = pv.Plotter(off_screen=off_screen)
-        if kwargs.pop('show_axes', True):
+        if kwargs.pop("show_axes", True):
             plotter.add_axes()
 
-        if 'rng' not in kwargs:
+        if "rng" not in kwargs:
             smax = np.abs(scalars).max()
-            if comp == 'norm':
-                kwargs['rng'] = [0, smax]
+            if comp == "norm":
+                kwargs["rng"] = [0, smax]
             else:
-                kwargs['rng'] = [-smax, smax]
+                kwargs["rng"] = [-smax, smax]
 
-        background = kwargs.pop('background', None)
+        background = kwargs.pop("background", None)
         if background:
             plotter.set_background(background)
 
-        plotter.add_mesh(plot_mesh,
-                         scalars=np.real(scalars),
-                         **kwargs)
+        plotter.add_mesh(plot_mesh, scalars=np.real(scalars), **kwargs)
 
         # setup text
-        plotter.add_text(' ', font_size=20, position=[0, 0], color=text_color)
+        plotter.add_text(" ", font_size=20, position=[0, 0], color=text_color)
 
         if cpos:
             plotter.camera_position = cpos
 
         if movie_filename:
-            if movie_filename.strip()[-3:] == 'gif':
+            movie_filename = str(movie_filename)
+            if movie_filename.strip()[-3:] == "gif":
                 plotter.open_gif(movie_filename)
             else:
                 plotter.open_movie(movie_filename)
 
         self._animating = True
+
         def q_callback():
             """exit when user wants to leave"""
             self._animating = False
@@ -1652,26 +1717,28 @@ class CyclicResult(Result):
         plotter.add_key_event("q", q_callback)
 
         # run until q is pressed
-        plotter.show(interactive=False, auto_close=False,
-                     interactive_update=not off_screen)
+        plotter.show(
+            interactive=False, auto_close=False, interactive_update=not off_screen
+        )
 
         first_loop = True
         while self._animating:
-            for angle in np.linspace(0, np.pi*2, nangles):
-                padj = 1*np.cos(angle) - 1j*np.sin(angle)
-                complex_disp_adj = np.real(complex_disp*padj)
+            for angle in np.linspace(0, np.pi * 2, n_frames):
+                padj = 1 * np.cos(angle) - 1j * np.sin(angle)
+                complex_disp_adj = np.real(complex_disp * padj)
 
                 if axis is not None:
                     scalars = complex_disp_adj[:, axis]
                 else:
-                    scalars = (complex_disp_adj*complex_disp_adj).sum(1)**0.5
+                    scalars = (complex_disp_adj * complex_disp_adj).sum(1) ** 0.5
 
                 plotter.update_scalars(scalars, render=False)
                 plot_mesh.points[:] = orig_pt + complex_disp_adj
 
                 if add_text:
-                    plotter.textActor.SetInput('%s\nPhase %.1f Degrees' %
-                                               (result_info, (angle*180/np.pi)))
+                    plotter.textActor.SetInput(
+                        "%s\nPhase %.1f Degrees" % (result_info, (angle * 180 / np.pi))
+                    )
 
                 plotter.update(1, force_redraw=True)
                 if not self._animating:
@@ -1694,16 +1761,16 @@ class CyclicResult(Result):
         return self.animate_nodal_solution(*args, **kwargs)
 
     def _gen_full_rotor(self):
-        """ Create full rotor vtk unstructured grid """
+        """Create full rotor vtk unstructured grid"""
         grid = self._mas_grid.copy()
         # transform to standard coordinate system
-        cs_cord = self._resultheader['csCord']
+        cs_cord = self._resultheader["csCord"]
         if cs_cord > 1:
             matrix = self.cs_4x4(cs_cord, as_vtk_matrix=True)
             grid.transform(matrix)
 
         # consider forcing low and high to be exact
-        # self._mas_grid.point_arrays['CYCLIC_M01H'] --> rotate and match
+        # self._mas_grid.point_data['CYCLIC_M01H'] --> rotate and match
 
         vtkappend = vtkAppendFilter()
         # vtkappend.MergePointsOn()
@@ -1714,8 +1781,8 @@ class CyclicResult(Result):
             sector = grid.copy()
             sector_id = np.empty(grid.n_points)
             sector_id[:] = i
-            sector.point_arrays['sector_id'] = sector_id
-            sector.rotate_z(rang * i)
+            sector.point_data["sector_id"] = sector_id
+            sector.rotate_z(rang * i, inplace=True)
             vtkappend.AddInputData(sector)
 
         vtkappend.Update()
@@ -1734,17 +1801,21 @@ class CyclicResult(Result):
             self._rotor_cache = self._gen_full_rotor()
         return self._rotor_cache
 
-    def _plot_cyclic_point_scalars(self, scalars, rnum,
-                                   show_displacement=False,
-                                   displacement_factor=1,
-                                   overlay_wireframe=False,
-                                   add_text=True,
-                                   node_components=None,
-                                   element_components=None,
-                                   sel_type_all=True,
-                                   phase=None,
-                                   treat_nan_as_zero=True,
-                                   **kwargs):
+    def _plot_cyclic_point_scalars(
+        self,
+        scalars,
+        rnum,
+        show_displacement=False,
+        displacement_factor=1,
+        overlay_wireframe=False,
+        add_text=True,
+        node_components=None,
+        element_components=None,
+        sel_type_all=True,
+        phase=None,
+        treat_nan_as_zero=False,
+        **kwargs,
+    ):
         """Plot point scalars on active mesh.
 
         Parameters
@@ -1788,8 +1859,7 @@ class CyclicResult(Result):
         # extract a sub-component of the grid if requested
         grid = self._mas_grid
         if node_components:
-            grid, ind = self._extract_node_components(node_components,
-                                                      sel_type_all)
+            grid, ind = self._extract_node_components(node_components, sel_type_all)
             if scalars is not None:
                 scalars = scalars[:, ind]
 
@@ -1802,25 +1872,25 @@ class CyclicResult(Result):
             scalars[np.isnan(scalars)] = 0
 
         # must be removed before add_mesh **kwargs
-        window_size = kwargs.pop('window_size', None)
-        full_screen = kwargs.pop('full_screen', False)
-        screenshot = kwargs.pop('screenshot', None)
-        text_color = kwargs.pop('text_color', None)
-        kwargs.setdefault('cmap', 'jet')
+        window_size = kwargs.pop("window_size", None)
+        full_screen = kwargs.pop("full_screen", False)
+        screenshot = kwargs.pop("screenshot", None)
+        text_color = kwargs.pop("text_color", None)
+        kwargs.setdefault("cmap", "jet")
         if scalars is not None:
-            kwargs.setdefault('rng', [np.nanmin(scalars), np.nanmax(scalars)])
+            kwargs.setdefault("rng", [np.nanmin(scalars), np.nanmax(scalars)])
 
         # Plot off screen when not interactive
-        off_screen = kwargs.pop('off_screen', None)
+        off_screen = kwargs.pop("off_screen", None)
         plotter = pv.Plotter(off_screen=off_screen)
 
         # various plotting properties that must be removed before add_mesh
-        if kwargs.pop('show_axes', True):
+        if kwargs.pop("show_axes", True):
             plotter.add_axes()
-        plotter.background_color = kwargs.pop('background', None)
-        cpos = kwargs.pop('cpos', None)
+        plotter.background_color = kwargs.pop("background", None)
+        cpos = kwargs.pop("cpos", None)
 
-        cs_cord = self._resultheader['csCord']
+        cs_cord = self._resultheader["csCord"]
         if cs_cord > 1:
             matrix = self.cs_4x4(cs_cord, as_vtk_matrix=True)
             i_matrix = self.cs_4x4(cs_cord, as_vtk_matrix=True)
@@ -1832,14 +1902,18 @@ class CyclicResult(Result):
         if overlay_wireframe:
             rang = 360.0 / self.n_sector
             for i in range(self.n_sector):
-                actor = plotter.add_mesh(grid.copy(False), color='w',
-                                         style='wireframe',
-                                         opacity=0.5, **kwargs)
+                actor = plotter.add_mesh(
+                    grid.copy(False),
+                    color="w",
+                    style="wireframe",
+                    opacity=0.5,
+                    **kwargs,
+                )
 
                 # transform to standard position, rotate about Z axis,
                 # transform back
                 transform = vtkTransform()
-                transform.RotateZ(rang*i)
+                transform.RotateZ(rang * i)
                 transform.Update()
                 rot_matrix = transform.GetMatrix()
 
@@ -1866,13 +1940,11 @@ class CyclicResult(Result):
 
             rotor = self.full_rotor.copy()
             rotor.points += disp
-            actor = plotter.add_mesh(rotor,
-                                     scalars=scalars.reshape(-1, 3),
-                                     **kwargs)
+            actor = plotter.add_mesh(rotor, scalars=scalars.reshape(-1, 3), **kwargs)
 
         else:
             surf_sector = grid.extract_surface()
-            ind = surf_sector.point_arrays['vtkOriginalPointIds']
+            ind = surf_sector.point_data["vtkOriginalPointIds"]
             rang = 360.0 / self.n_sector
             for i in range(self.n_sector):
                 if scalars is not None:
@@ -1880,9 +1952,9 @@ class CyclicResult(Result):
                 else:
                     sector_scalars = None
 
-                actor = plotter.add_mesh(surf_sector.copy(False),
-                                         scalars=sector_scalars,
-                                         **kwargs)
+                actor = plotter.add_mesh(
+                    surf_sector.copy(False), scalars=sector_scalars, **kwargs
+                )
 
                 # NAN/missing data are white
                 plotter.mapper.GetLookupTable().SetNanColor(1, 1, 1, 1)
@@ -1890,7 +1962,7 @@ class CyclicResult(Result):
                 # transform to standard position, rotate about Z axis,
                 # transform back
                 transform = vtkTransform()
-                transform.RotateZ(rang*i)
+                transform.RotateZ(rang * i)
                 transform.Update()
                 rot_matrix = transform.GetMatrix()
 
@@ -1904,21 +1976,21 @@ class CyclicResult(Result):
 
         # add table
         if isinstance(add_text, str):
-            plotter.add_text(add_text, font_size=20, position=[0, 0],
-                             color=text_color)
+            plotter.add_text(add_text, font_size=20, position=[0, 0], color=text_color)
         elif add_text:
             rnum = self.parse_step_substep(rnum)
-            plotter.add_text(self.text_result_table(rnum), font_size=20,
-                             color=text_color)
+            plotter.add_text(
+                self.text_result_table(rnum), font_size=20, color=text_color
+            )
 
         # must set camera position at the ended
         if cpos is not None:
             plotter.camera_position = cpos
 
         if screenshot:
-            cpos = plotter.show(auto_close=False,
-                                window_size=window_size,
-                                full_screen=full_screen)
+            cpos = plotter.show(
+                auto_close=False, window_size=window_size, full_screen=full_screen
+            )
             plotter.screenshot(screenshot)
             plotter.close()
         else:
