@@ -1,26 +1,35 @@
-"""Module to read ANSYS ASCII block formatted CDB files"""
-import io
-import os
-import logging
+"""Module to read ANSYS ASCII block formatted CDB files."""
 from functools import wraps
+import io
+import logging
+import os
+import pathlib
 
 import numpy as np
-from vtk import (VTK_TETRA, VTK_QUADRATIC_TETRA, VTK_PYRAMID,
-                 VTK_QUADRATIC_PYRAMID, VTK_WEDGE, VTK_QUADRATIC_WEDGE,
-                 VTK_HEXAHEDRON, VTK_QUADRATIC_HEXAHEDRON, VTK_TRIANGLE, VTK_QUAD,
-                 VTK_QUADRATIC_TRIANGLE, VTK_QUADRATIC_QUAD)
-import vtk
 import pyvista as pv
+from pyvista import _vtk as vtk
+from pyvista._vtk import (
+    VTK_HEXAHEDRON,
+    VTK_PYRAMID,
+    VTK_QUAD,
+    VTK_QUADRATIC_HEXAHEDRON,
+    VTK_QUADRATIC_PYRAMID,
+    VTK_QUADRATIC_TETRA,
+    VTK_QUADRATIC_WEDGE,
+    VTK_TETRA,
+    VTK_TRIANGLE,
+    VTK_WEDGE,
+)
 
-from ansys.mapdl.reader import _reader, _archive
-from ansys.mapdl.reader.misc import vtk_cell_info, chunks
-from ansys.mapdl.reader.mesh import Mesh
+VTK_VOXEL = 11
+
+from ansys.mapdl.reader import _archive, _reader
 from ansys.mapdl.reader.cell_quality import quality
-
-VTK9 = vtk.vtkVersion().GetVTKMajorVersion() >= 9
+from ansys.mapdl.reader.mesh import Mesh
+from ansys.mapdl.reader.misc import vtk_cell_info
 
 log = logging.getLogger(__name__)
-log.setLevel('CRITICAL')
+log.setLevel("CRITICAL")
 
 
 class Archive(Mesh):
@@ -35,7 +44,7 @@ class Archive(Mesh):
 
     Parameters
     ----------
-    filename : string
+    filename : string, pathlib.Path
         Filename of block formatted cdb file
 
     read_parameters : bool, optional
@@ -67,6 +76,9 @@ class Archive(Mesh):
 
     name : str, optional
         Internally used parameter used to have a custom ``__repr__``.
+
+    read_eblock : bool, default: True
+        Read the element block.
 
     Examples
     --------
@@ -105,44 +117,69 @@ class Archive(Mesh):
     records.  These additional records will be ignored.
     """
 
-    def __init__(self, filename, read_parameters=False,
-                 parse_vtk=True, force_linear=False,
-                 allowable_types=None, null_unallowed=False,
-                 verbose=False, name=''):
+    def __init__(
+        self,
+        filename,
+        read_parameters=False,
+        parse_vtk=True,
+        force_linear=False,
+        allowable_types=None,
+        null_unallowed=False,
+        verbose=False,
+        name="",
+        read_eblock=True,
+    ):
         """Initializes an instance of the archive class."""
         self._read_parameters = read_parameters
-        self._filename = filename
+        self._filename = pathlib.Path(filename)
         self._name = name
-        self._raw = _reader.read(filename, read_parameters=read_parameters,
-                                 debug=verbose)
-        super().__init__(self._raw['nnum'],
-                         self._raw['nodes'],
-                         self._raw['elem'],
-                         self._raw['elem_off'],
-                         self._raw['ekey'],
-                         node_comps=self._raw['node_comps'],
-                         elem_comps=self._raw['elem_comps'],
-                         rdat=self._raw['rdat'],
-                         rnum=self._raw['rnum'],
-                         keyopt=self._raw['keyopt'])
+        self._raw = _reader.read(
+            self.filename,
+            read_parameters=read_parameters,
+            debug=verbose,
+            read_eblock=read_eblock,
+        )
+        super().__init__(
+            self._raw["nnum"],
+            self._raw["nodes"],
+            self._raw["elem"],
+            self._raw["elem_off"],
+            self._raw["ekey"],
+            node_comps=self._raw["node_comps"],
+            elem_comps=self._raw["elem_comps"],
+            rdat=self._raw["rdat"],
+            rnum=self._raw["rnum"],
+            keyopt=self._raw["keyopt"],
+        )
 
         self._allowable_types = allowable_types
         self._force_linear = force_linear
         self._null_unallowed = null_unallowed
 
         if parse_vtk:
-            self._grid = self._parse_vtk(allowable_types,
-                                         force_linear, null_unallowed)
+            self._grid = self._parse_vtk(allowable_types, force_linear, null_unallowed)
+
+    @property
+    def filename(self) -> str:
+        """String form of the filename. This property is read-only."""
+        return str(self._filename)
+
+    @property
+    def pathlib_filename(self) -> pathlib.Path:
+        """Return the ``pathlib.Path`` version of the filename. This property can not be set."""
+        return self._filename
 
     @property
     def raw(self):  # pragma: no cover
-        raise AttributeError('The `raw` attribute has been depreciated.  Access'
-                             ' the values directly from the archive object.\n\n'
-                             '    Instead of:\n'
-                             '    archive.raw["nodes"]\n'
-                             '    \n'
-                             '    Use\n'
-                             '    print(archive.nodes)')
+        raise AttributeError(
+            "The `raw` attribute has been depreciated.  Access"
+            " the values directly from the archive object.\n\n"
+            "    Instead of:\n"
+            '    archive.raw["nodes"]\n'
+            "    \n"
+            "    Use\n"
+            "    print(archive.nodes)"
+        )
 
     @property
     def parameters(self):
@@ -158,22 +195,24 @@ class Archive(Mesh):
         {}
         """
         if not self._read_parameters:
-            raise AttributeError('No parameters read.  Read the archive again '
-                                 ' with ``read_parameters=True``')
-        return self._raw['parameters']
+            raise AttributeError(
+                "No parameters read.  Read the archive again "
+                " with ``read_parameters=True``"
+            )
+        return self._raw["parameters"]
 
     def __repr__(self):
         if self._name:
-            txt = 'MAPDL %s\n' % self._name
+            txt = "MAPDL %s\n" % self._name
         else:
             basename = os.path.basename(self._filename)
-            txt = 'ANSYS Archive File %s\n' % basename
+            txt = "ANSYS Archive File %s\n" % basename
 
-        txt += '  Number of Nodes:              %d\n' % len(self.nnum)
-        txt += '  Number of Elements:           %d\n' % len(self.enum)
-        txt += '  Number of Element Types:      %d\n' % len(self.ekey)
-        txt += '  Number of Node Components:    %d\n' % len(self.node_components)
-        txt += '  Number of Element Components: %d\n' % len(self.element_components)
+        txt += "  Number of Nodes:              %d\n" % len(self.nnum)
+        txt += "  Number of Elements:           %d\n" % len(self.enum)
+        txt += "  Number of Element Types:      %d\n" % len(self.ekey)
+        txt += "  Number of Node Components:    %d\n" % len(self.node_components)
+        txt += "  Number of Element Components: %d\n" % len(self.element_components)
         return txt
 
     @property
@@ -195,9 +234,9 @@ class Archive(Mesh):
           N Arrays:     13
         """
         if self._grid is None:  # parse the grid using the cached parameters
-            self._grid = self._parse_vtk(self._allowable_types,
-                                         self._force_linear,
-                                         self._null_unallowed)
+            self._grid = self._parse_vtk(
+                self._allowable_types, self._force_linear, self._null_unallowed
+            )
         return self._grid
 
     @property
@@ -218,55 +257,69 @@ class Archive(Mesh):
                1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.])
         """
         if self._grid is None:  # pragma: no cover
-            raise AttributeError('Archive must be parsed as a vtk grid.\n'
-                                 'Set `parse_vtk=True`')
+            raise AttributeError(
+                "Archive must be parsed as a vtk grid.\n" "Set `parse_vtk=True`"
+            )
         return quality(self._grid)
 
     @wraps(pv.plot)
     def plot(self, *args, **kwargs):
         """Plot the mesh"""
         if self._grid is None:  # pragma: no cover
-            raise AttributeError('Archive must be parsed as a vtk grid.\n'
-                                 'Set `parse_vtk=True`')
-        kwargs.setdefault('color', 'w')
-        kwargs.setdefault('show_edges', True)
+            raise AttributeError(
+                "Archive must be parsed as a vtk grid.\n" "Set `parse_vtk=True`"
+            )
+        kwargs.setdefault("color", "w")
+        kwargs.setdefault("show_edges", True)
         self.grid.plot(*args, **kwargs)
 
 
-def save_as_archive(filename, grid, mtype_start=1, etype_start=1,
-                    real_constant_start=1, mode='w',
-                    enum_start=1, nnum_start=1,
-                    include_etype_header=True, reset_etype=False,
-                    allow_missing=True, include_surface_elements=True,
-                    include_solid_elements=True, include_components=True):
-    """Writes FEM as an ANSYS APDL archive file.  This function
-    supports the following element types:
+def save_as_archive(
+    filename,
+    grid,
+    mtype_start=1,
+    etype_start=1,
+    real_constant_start=1,
+    mode="w",
+    enum_start=1,
+    nnum_start=1,
+    include_etype_header=True,
+    reset_etype=False,
+    allow_missing=True,
+    include_surface_elements=True,
+    include_solid_elements=True,
+    include_components=True,
+    exclude_missing=False,
+):
+    """Writes FEM as an ANSYS APDL archive file.
 
-        - ``vtk.VTK_TETRA``
-        - ``vtk.VTK_QUADRATIC_TETRA``
-        - ``vtk.VTK_PYRAMID``
-        - ``vtk.VTK_QUADRATIC_PYRAMID``
-        - ``vtk.VTK_WEDGE``
-        - ``vtk.VTK_QUADRATIC_WEDGE``
+    This function supports the following element types:
+
         - ``vtk.VTK_HEXAHEDRON``
+        - ``vtk.VTK_PYRAMID``
         - ``vtk.VTK_QUADRATIC_HEXAHEDRON``
-        - ``vtk.VTK_TRIANGLE``
+        - ``vtk.VTK_QUADRATIC_PYRAMID``
+        - ``vtk.VTK_QUADRATIC_TETRA``
+        - ``vtk.VTK_QUADRATIC_WEDGE``
         - ``vtk.VTK_QUAD``
+        - ``vtk.VTK_TETRA``
+        - ``vtk.VTK_TRIANGLE``
+        - ``vtk.VTK_VOXEL``
+        - ``vtk.VTK_WEDGE``
 
     Will automatically renumber nodes and elements if the FEM does not
     contain ANSYS node or element numbers.  Node numbers are stored as
-    a point array "ansys_node_num", and cell numbers are stored as
-    cell array "ansys_elem_num".
+    a point array ``"ansys_node_num"``, and cell numbers are stored as
+    cell array ``"ansys_elem_num"``.
 
     Parameters
     ----------
-    filename : str
+    filename : str, pathlib.Path
        Filename to write archive file.
 
-    grid : vtk.UnstructuredGrid
-        VTK UnstructuredGrid to convert to an APDL archive file.
-        PolyData will automatically be converted to an unstructured
-        mesh.
+    grid : pyvista.DataSet
+        Any :class:`pyvista.DataSet` that can be cast to a
+        :class:`pyvista.UnstructuredGrid`.
 
     mtype_start : int, optional
         Material number to assign to elements.  Can be set manually by
@@ -316,9 +369,15 @@ def save_as_archive(filename, grid, mtype_start=1, etype_start=1,
         Writes note components to file.  Node components must be
         stored within the unstructured grid as uint8 or bool arrays.
 
+    exclude_missing : bool, default: False
+        When ``allow_missing=True``, write ``0`` instead of renumbering
+        nodes. This allows you to exclude midside nodes for certain element
+        types (e.g. ``SOLID186``). Missing midside nodes are identified as
+        ``-1`` in the ``"ansys_node_num"`` array.
+
     Examples
     --------
-    Write a VTK of pyvista UnstructuredGrid to archive.cdb
+    Write a ``pyvista.UnstructuredGrid`` to ``"archive.cdb"``.
 
     >>> from ansys.mapdl import reader as pymapdl_reader
     >>> from pyvista import examples
@@ -326,137 +385,179 @@ def save_as_archive(filename, grid, mtype_start=1, etype_start=1,
     >>> pymapdl_reader.save_as_archive('archive.cdb', grid)
 
     """
-    if isinstance(grid, pv.PolyData):
+    if hasattr(grid, "cast_to_unstructured_grid"):
         grid = grid.cast_to_unstructured_grid()
 
     if not isinstance(grid, vtk.vtkUnstructuredGrid):
-        raise TypeError('``grid`` argument must be an UnstructuredGrid')
+        raise TypeError("``grid`` argument must be an UnstructuredGrid")
 
     allowable = []
     if include_solid_elements:
-        allowable.extend([VTK_TETRA,
-                          VTK_QUADRATIC_TETRA,
-                          VTK_PYRAMID,
-                          VTK_QUADRATIC_PYRAMID,
-                          VTK_WEDGE,
-                          VTK_QUADRATIC_WEDGE,
-                          VTK_HEXAHEDRON,
-                          VTK_QUADRATIC_HEXAHEDRON])
+        allowable.extend(
+            [
+                VTK_VOXEL,
+                VTK_TETRA,
+                VTK_QUADRATIC_TETRA,
+                VTK_PYRAMID,
+                VTK_QUADRATIC_PYRAMID,
+                VTK_WEDGE,
+                VTK_QUADRATIC_WEDGE,
+                VTK_HEXAHEDRON,
+                VTK_QUADRATIC_HEXAHEDRON,
+            ]
+        )
 
     if include_surface_elements:
-        allowable.extend([VTK_TRIANGLE,
-                          VTK_QUAD])
-                          # VTK_QUADRATIC_TRIANGLE,
-                          # VTK_QUADRATIC_QUAD
+        allowable.extend([VTK_TRIANGLE, VTK_QUAD])
+        # VTK_QUADRATIC_TRIANGLE,
+        # VTK_QUADRATIC_QUAD
 
     # extract allowable cell types
     mask = np.in1d(grid.celltypes, allowable)
+    if not mask.any():
+        ucelltypes = np.unique(grid.celltypes)
+        allowable.sort()
+        raise RuntimeError(
+            f"`grid` contains no allowable cell types. Contains types {ucelltypes} "
+            f"and only {allowable} are allowed.\n\n"
+            "See https://vtk.org/doc/nightly/html/vtkCellType_8h_source.html "
+            "for more details."
+        )
     grid = grid.extract_cells(mask)
 
-    header = '/PREP7\n'
+    header = "/PREP7\n"
 
     # node numbers
-    if 'ansys_node_num' in grid.point_arrays:
-        nodenum = grid.point_arrays['ansys_node_num']
+    if "ansys_node_num" in grid.point_data:
+        nodenum = grid.point_data["ansys_node_num"]
     else:
-        log.info('No ANSYS node numbers set in input. Adding default range')
+        log.info("No ANSYS node numbers set in input. Adding default range")
         nodenum = np.arange(1, grid.number_of_points + 1, dtype=np.int32)
 
-    if np.any(nodenum == -1):
+    missing_mask = nodenum == -1
+    if np.any(missing_mask):
         if not allow_missing:
             raise Exception('Missing node numbers.  Exiting due "allow_missing=False"')
-        start_num = nodenum.max() + 1
-        if nnum_start > start_num:
-            start_num = nnum_start
-        nadd = np.sum(nodenum == -1)
-        end_num = start_num + nadd
-        log.info('FEM missing some node numbers.  Adding node numbering '
-                 'from %d to %d', start_num, end_num)
-        nodenum[nodenum == -1] = np.arange(start_num, end_num, dtype=np.int32)
+        elif exclude_missing:
+            log.info("Excluding missing nodes from archive file.")
+            nodenum = nodenum.copy()
+            nodenum[missing_mask] = 0
+        else:
+            start_num = nodenum.max() + 1
+            if nnum_start > start_num:
+                start_num = nnum_start
+            nadd = np.sum(nodenum == -1)
+            end_num = start_num + nadd
+            log.info(
+                "FEM missing some node numbers.  Adding node numbering "
+                "from %d to %d",
+                start_num,
+                end_num,
+            )
+            nodenum[missing_mask] = np.arange(start_num, end_num, dtype=np.int32)
 
     # element block
     ncells = grid.number_of_cells
-    if 'ansys_elem_num' in grid.cell_arrays:
-        enum = grid.cell_arrays['ansys_elem_num']
+    if "ansys_elem_num" in grid.cell_data:
+        enum = grid.cell_data["ansys_elem_num"]
     else:
         if not allow_missing:
             raise Exception('Missing node numbers.  Exiting due "allow_missing=False"')
-        log.info('No ANSYS element numbers set in input.  ' +
-                 'Adding default range starting from %d' % enum_start)
+        log.info(
+            "No ANSYS element numbers set in input. "
+            "Adding default range starting from %d",
+            enum_start,
+        )
         enum = np.arange(1, ncells + 1, dtype=np.int32)
 
     if np.any(enum == -1):
         if not allow_missing:
-            raise Exception('-1 encountered in "ansys_elem_num".\n'
-                            'Exiting due "allow_missing=False"')
+            raise Exception(
+                '-1 encountered in "ansys_elem_num".\n'
+                'Exiting due "allow_missing=False"'
+            )
 
         start_num = enum.max() + 1
         if enum_start > start_num:
             start_num = enum_start
         nadd = np.sum(enum == -1)
         end_num = start_num + nadd
-        log.info('FEM missing some cell numbers.  Adding numbering ' +
-                 'from %d to %d' % (start_num, end_num))
+        log.info(
+            "FEM missing some cell numbers.  Adding numbering " "from %d to %d",
+            start_num,
+            end_num,
+        )
         enum[enum == -1] = np.arange(start_num, end_num, dtype=np.int32)
 
     # material type
-    if 'ansys_material_type' in grid.cell_arrays:
-        mtype = grid.cell_arrays['ansys_material_type']
+    if "ansys_material_type" in grid.cell_data:
+        mtype = grid.cell_data["ansys_material_type"]
     else:
-        log.info('No ANSYS element numbers set in input.  ' +
-                 'Adding default range starting from %d', mtype_start)
+        log.info(
+            "No ANSYS element numbers set in input.  "
+            "Adding default range starting from %d",
+            mtype_start,
+        )
         mtype = np.arange(1, ncells + 1, dtype=np.int32)
 
     if np.any(mtype == -1):
-        log.info('FEM missing some material type numbers.  Adding...')
+        log.info("FEM missing some material type numbers.  Adding...")
         mtype[mtype == -1] = mtype_start
 
     # real constant
-    if 'ansys_real_constant' in grid.cell_arrays:
-        rcon = grid.cell_arrays['ansys_real_constant']
+    if "ansys_real_constant" in grid.cell_data:
+        rcon = grid.cell_data["ansys_real_constant"]
     else:
-        log.info('No ANSYS element numbers set in input.  ' +
-                 'Adding default range starting from %d', real_constant_start)
+        log.info(
+            "No ANSYS element numbers set in input.  "
+            + "Adding default range starting from %d",
+            real_constant_start,
+        )
         rcon = np.arange(1, ncells + 1, dtype=np.int32)
 
     if np.any(rcon == -1):
-        log.info('FEM missing some material type numbers.  Adding...')
+        log.info("FEM missing some material type numbers.  Adding...")
         rcon[rcon == -1] = real_constant_start
 
     # element type
     invalid = False
-    if 'ansys_etype' in grid.cell_arrays and not reset_etype:
+    if "ansys_etype" in grid.cell_data and not reset_etype:
         missing = False
-        typenum = grid.cell_arrays['ansys_elem_type_num']
-        etype = grid.cell_arrays['ansys_etype']
+        typenum = grid.cell_data["ansys_elem_type_num"]
+        etype = grid.cell_data["ansys_etype"]
         if np.any(etype == -1):
-            log.warning('Some elements are missing element type numbers.')
+            log.warning("Some elements are missing element type numbers.")
             invalid = True
 
         if include_etype_header and not invalid:
             _, ind = np.unique(etype, return_index=True)
             for idx in ind:
-                header += 'ET, %d, %d\n' % (etype[idx], typenum[idx])
+                header += "ET, %d, %d\n" % (etype[idx], typenum[idx])
     else:
         missing = True
 
     # check if valid
     if not missing:
         mask = grid.celltypes < 20
-        if np.any(grid.cell_arrays['ansys_elem_type_num'][mask] == 186):
+        if np.any(grid.cell_data["ansys_elem_type_num"][mask] == 186):
             invalid = True
-            log.warning('Invalid ANSYS element types.')
+            log.warning("Invalid ANSYS element types.")
 
     if invalid or missing:
         if not allow_missing:
-            raise Exception('Invalid or missing data in "ansys_elem_type_num"' +
-                            ' or "ansys_etype".  Exiting due "allow_missing=False"')
+            raise Exception(
+                'Invalid or missing data in "ansys_elem_type_num"'
+                + ' or "ansys_etype".  Exiting due "allow_missing=False"'
+            )
 
-        log.info('No ANSYS element type or invalid data input.  ' +
-                 'Adding default range starting from %d' % etype_start)
+        log.info(
+            "No ANSYS element type or invalid data input.  "
+            + "Adding default range starting from %d" % etype_start
+        )
 
         etype = np.empty(grid.number_of_cells, np.int32)
         etype_185 = etype_start + 2
+        etype[grid.celltypes == VTK_VOXEL] = etype_185
         etype[grid.celltypes == VTK_TETRA] = etype_185
         etype[grid.celltypes == VTK_HEXAHEDRON] = etype_185
         etype[grid.celltypes == VTK_WEDGE] = etype_185
@@ -481,10 +582,10 @@ def save_as_archive(filename, grid, mtype_start=1, etype_start=1,
         typenum[etype == etype_187] = 187
         typenum[etype == etype_181] = 181
 
-        header += 'ET, %d, 185\n' % etype_185
-        header += 'ET, %d, 186\n' % etype_186
-        header += 'ET, %d, 187\n' % etype_187
-        header += 'ET, %d, 181\n' % etype_181
+        header += "ET, %d, 185\n" % etype_185
+        header += "ET, %d, 186\n" % etype_186
+        header += "ET, %d, 187\n" % etype_187
+        header += "ET, %d, 181\n" % etype_181
 
     # number of nodes written per element
     elem_nnodes = np.empty(etype.size, np.int32)
@@ -493,39 +594,69 @@ def save_as_archive(filename, grid, mtype_start=1, etype_start=1,
     elem_nnodes[typenum == 186] = 20
     elem_nnodes[typenum == 187] = 10
 
+    if not reset_etype:
+        unsup = np.setdiff1d(typenum, [181, 185, 186, 187])
+        if unsup.any():
+            raise RuntimeError(
+                f"Unsupported element types {unsup}. Either set ``reset_etype=True``"
+                " or remove (or relabel) the unsupported element types."
+            )
+
+    # edge case where element types are unsupported
+
     # write the EBLOCK
     with open(str(filename), mode) as f:
         f.write(header)
 
     if not isinstance(filename, str):
         filename = str(filename)
-    write_nblock(filename, nodenum, grid.points, mode='a')
+
+    if exclude_missing:
+        log.info("Excluding missing nodes from archive file.")
+        write_nblock(
+            filename,
+            nodenum[~missing_mask],
+            grid.points[~missing_mask],
+            mode="a",
+        )
+    else:
+        write_nblock(filename, nodenum, grid.points, mode="a")
 
     # write remainder of eblock
     cells, offset = vtk_cell_info(grid, shift_offset=False)
-    _write_eblock(filename, enum, etype, mtype, rcon, elem_nnodes,
-                  cells, offset, grid.celltypes, typenum, nodenum,
-                  mode='a')
+    _write_eblock(
+        filename,
+        enum,
+        etype,
+        mtype,
+        rcon,
+        elem_nnodes,
+        cells,
+        offset,
+        grid.celltypes,
+        typenum,
+        nodenum,
+        mode="a",
+    )
 
     if include_components:
-        with open(filename, 'a') as fid:
-
+        with open(filename, "a") as fid:
             # write node components
-            for node_key in grid.point_arrays:
-                arr = grid.point_arrays[node_key]
-                if arr.dtype in [np.uint8, np.bool]:
-                    items = nodenum[arr.view(np.bool)]
-                    write_cmblock(fid, items, node_key, 'NODE')
+            for node_key in grid.point_data:
+                arr = grid.point_data[node_key]
+                if arr.dtype in [np.uint8, np.bool_]:
+                    items = nodenum[arr.view(np.bool_)]
+                    write_cmblock(fid, items, node_key, "NODE")
 
             # write element components
-            for node_key in grid.cell_arrays:
-                arr = grid.cell_arrays[node_key]
-                if arr.dtype in [np.uint8, np.bool]:
-                    items = enum[arr.view(np.bool)]
-                    write_cmblock(fid, items, node_key, 'ELEMENT')
+            for node_key in grid.cell_data:
+                arr = grid.cell_data[node_key]
+                if arr.dtype in [np.uint8, np.bool_]:
+                    items = enum[arr.view(np.bool_)]
+                    write_cmblock(fid, items, node_key, "ELEMENT")
 
 
-def write_nblock(filename, node_id, pos, angles=None, mode='w'):
+def write_nblock(filename, node_id, pos, angles=None, mode="w"):
     """Writes nodes and node angles to file.
 
     Parameters
@@ -545,12 +676,11 @@ def write_nblock(filename, node_id, pos, angles=None, mode='w'):
     mode : str, optional
         Write mode.  Default ``'w'``.
     """
-    assert pos.ndim == 2 and pos.shape[1] == 3, 'Invalid position array'
+    assert pos.ndim == 2 and pos.shape[1] == 3, "Invalid position array"
     if angles is not None:
-        assert angles.ndim == 2 and angles.shape[1] == 3, 'Invalid angle array'
+        assert angles.ndim == 2 and angles.shape[1] == 3, "Invalid angle array"
 
-    if node_id.dtype != np.int32:
-        node_id = node_id.astype(np.int32)
+    node_id = node_id.astype(np.int32, copy=False)
 
     # node array must be sorted
     # note, this is sort check is most suited for pre-sorted arrays
@@ -562,66 +692,55 @@ def write_nblock(filename, node_id, pos, angles=None, mode='w'):
 
     if angles is not None:
         if pos.dtype == np.float32:
-            if angles.dtype != pos.dtype:
-                angles = angles.astype(pos.dtype)
-            _archive.py_write_nblock_float(filename,
-                                           node_id,
-                                           node_id[-1],
-                                           pos,
-                                           angles,
-                                           mode)
+            angles = angles.astype(pos.dtype, copy=False)
+            _archive.py_write_nblock_float(
+                filename, node_id, node_id[-1], pos, angles, mode
+            )
         else:
-            _archive.py_write_nblock(filename,
-                                     node_id,
-                                     node_id[-1],
-                                     pos,
-                                     angles,
-                                     mode)
+            _archive.py_write_nblock(filename, node_id, node_id[-1], pos, angles, mode)
 
     else:
         if pos.dtype == np.float32:
-            _archive.py_write_nblock_float(filename,
-                                           node_id,
-                                           node_id[-1],
-                                           pos,
-                                           np.empty((0, 0), dtype=np.float32),
-                                           mode)
+            _archive.py_write_nblock_float(
+                filename,
+                node_id,
+                node_id[-1],
+                pos,
+                np.empty((0, 0), dtype=np.float32),
+                mode,
+            )
         else:
-            _archive.py_write_nblock(filename,
-                                     node_id,
-                                     node_id[-1],
-                                     pos,
-                                     np.empty((0, 0), dtype=np.float64),
-                                     mode)
+            _archive.py_write_nblock(
+                filename,
+                node_id,
+                node_id[-1],
+                pos,
+                np.empty((0, 0), dtype=np.float64),
+                mode,
+            )
 
 
-def write_cmblock(filename, items, comp_name, comp_type,
-                  digit_width=10, mode='w'):
+def write_cmblock(filename, items, comp_name, comp_type, digit_width=10, mode="w"):
     """Writes a component block (CMBLOCK) to a file.
 
     Parameters
     ----------
     filename : str or file handle
         File to write CMBLOCK component to.
-
     items : list or np.ndarray
         Element or node numbers to write.
-
     comp_name : str
         Name of the component
-
     comp_type : str
         Component type to write.  Should be either 'ELEMENT' or 'NODE'.
-
     digit_width : int, optional
         Default 10
-
     mode : str, optional
         Write mode.  Default ``'w'``.
     """
     comp_name = comp_name.upper()
     comp_type = comp_type.upper()
-    if comp_type.upper() not in ['ELEMENT', 'NODE']:
+    if comp_type.upper() not in ["ELEMENT", "NODE"]:
         raise ValueError("`comp_type` must be either 'ELEMENT' or 'NODE'")
 
     opened_file = False
@@ -640,28 +759,43 @@ def write_cmblock(filename, items, comp_name, comp_type,
     # consider cythonizing this in the future
     cmblock_items = _archive.cmblock_items_from_array(items)
     nitems = len(cmblock_items)
-    print(f'CMBLOCK,{comp_name},{comp_type},{nitems:8d}', file=fid)
-    print(f'(8i{digit_width})', file=fid)
-    digit_formatter = f'%{digit_width}d'
+    print(f"CMBLOCK,{comp_name},{comp_type},{nitems:8d}", file=fid)
+    print(f"(8i{digit_width})", file=fid)
+    digit_formatter = f"%{digit_width}d"
 
     # use np savetxt here as it's faster than looping through and
     # writing each line.
     # nearest multiple of 8
     up_to = len(cmblock_items) % 8
-    np.savetxt(fid, cmblock_items[:-up_to].reshape(-1, 8), digit_formatter*8)
+    if up_to:  # deal with the zero case
+        np.savetxt(fid, cmblock_items[:-up_to].reshape(-1, 8), digit_formatter * 8)
 
-    # write the final line
-    chunk = cmblock_items[-up_to:]
-    print(''.join([digit_formatter] * len(chunk)) % tuple(chunk), file=fid)
-    print('', file=fid)
+        # write the final line
+        chunk = cmblock_items[-up_to:]
+        print("".join([digit_formatter] * len(chunk)) % tuple(chunk), file=fid)
+    else:
+        np.savetxt(fid, cmblock_items.reshape(-1, 8), digit_formatter * 8)
+
+    print("", file=fid)
 
     if opened_file:
         fid.close()
 
 
-def _write_eblock(filename, elem_id, etype, mtype, rcon, elem_nnodes,
-                 cells, offset, celltypes, typenum,
-                 nodenum, mode='a'):
+def _write_eblock(
+    filename,
+    elem_id,
+    etype,
+    mtype,
+    rcon,
+    elem_nnodes,
+    cells,
+    offset,
+    celltypes,
+    typenum,
+    nodenum,
+    mode="a",
+):
     """Write EBLOCK to disk"""
     # perform type checking here
     if elem_id.dtype != np.int32:
@@ -685,16 +819,17 @@ def _write_eblock(filename, elem_id, etype, mtype, rcon, elem_nnodes,
     if nodenum.dtype != np.int32:
         nodenum = nodenum.astype(np.int32)
 
-    _archive.py_write_eblock(filename,
-                             elem_id,
-                             etype,
-                             mtype,
-                             rcon,
-                             elem_nnodes,
-                             cells,
-                             offset,
-                             celltypes,
-                             typenum,
-                             nodenum,
-                             VTK9,
-                             mode=mode)
+    _archive.py_write_eblock(
+        filename,
+        elem_id,
+        etype,
+        mtype,
+        rcon,
+        elem_nnodes,
+        cells,
+        offset,
+        celltypes,
+        typenum,
+        nodenum,
+        mode=mode,
+    )
