@@ -12,4 +12,4 @@ int write_nblock_float(FILE*, const int, const int, const int*, const float*,
                        const float*, int);
 int write_eblock(FILE*, const int, const int*, const int*, const int*,
                  const int*, const int*, const uint8_t*, const int64_t*,
-                 const int64_t*, const int*, const int*, const int);
+                 const int64_t*, const int*, const int*);
