@@ -1,27 +1,28 @@
 """Handle result files from a distributed MAPDL analysis"""
-from inspect import currentframe
-import glob
-import os
 from functools import wraps
+import glob
+from inspect import currentframe
+import os
+import pathlib
+from typing import Union
 
-import pyvista as pv
 import numpy as np
-from vtk import vtkAppendFilter
+import pyvista as pv
+from pyvista._vtk import vtkAppendFilter
 
-from ansys.mapdl.reader.misc import is_float, vtk_cell_info
-from ansys.mapdl.reader.mesh import Mesh
-from ansys.mapdl.reader.rst import Result, ELEMENT_INDEX_TABLE_KEYS
-from ansys.mapdl.reader.errors import NoDistributedFiles
-from ansys.mapdl.reader._binary_reader import (read_nodal_values_dist,
-                                               populate_surface_element_result)
+from ansys.mapdl.reader._binary_reader import read_nodal_values_dist
 from ansys.mapdl.reader._rst_keys import element_index_table_info
+from ansys.mapdl.reader.errors import NoDistributedFiles
+from ansys.mapdl.reader.mesh import Mesh
+from ansys.mapdl.reader.misc import is_float, vtk_cell_info
+from ansys.mapdl.reader.rst import ELEMENT_INDEX_TABLE_KEYS, Result
 
 
-def find_dis_files(main_file):
+def find_dis_files(main_file: Union[str, pathlib.Path]):
     """Find the individual distributed result files given a main result file"""
     basename = os.path.basename(main_file)
-    jobname = basename[:basename.rfind('0')]
-    ext = basename.split('.')[1]
+    jobname = basename[: basename.rfind("0")]
+    ext = basename.split(".")[1]
     dirname = os.path.dirname(main_file)
 
     if dirname is None:
@@ -29,8 +30,8 @@ def find_dis_files(main_file):
 
     # get all files matching the jobname
     filenames = {}
-    for filename in glob.glob('%s/%s*.%s' % (dirname, jobname, ext)):
-        index = os.path.basename(filename).replace(jobname, '')[:-4]
+    for filename in glob.glob("%s/%s*.%s" % (dirname, jobname, ext)):
+        index = os.path.basename(filename).replace(jobname, "")[:-4]
         if is_float(index):
             filenames[int(index)] = filename
 
@@ -38,8 +39,9 @@ def find_dis_files(main_file):
         raise NoDistributedFiles
 
     if max(filenames.keys()) + 1 != len(filenames):
-        raise FileNotFoundError('Unable to find all the result files of a '
-                                'distributed result')
+        raise FileNotFoundError(
+            "Unable to find all the result files of a " "distributed result"
+        )
 
     return filenames
 
@@ -65,14 +67,15 @@ class DistributedResult(Result):
 
         # Global number of nodes must not equal the number of nodes in this file
         if not self._main_result._is_distributed:
-            raise RuntimeError('Result file is not part of a distributed result')
+            raise RuntimeError("Result file is not part of a distributed result")
 
         if not self._main_result._is_main:  # pragma: no cover
-            raise RuntimeError('DistributedResult must be created from the main '
-                               'result file')
+            raise RuntimeError(
+                "DistributedResult must be created from the main " "result file"
+            )
 
         # load and verify
-        ptr = self._main_result._resultheader['ptrGNOD']
+        ptr = self._main_result._resultheader["ptrGNOD"]
         gl_nnum = self._main_result.read_record(ptr)
 
         mask = np.in1d(gl_nnum, self._main_result.mesh.nnum, assume_unique=True)
@@ -80,17 +83,19 @@ class DistributedResult(Result):
             result = Result(filenames[index])
             new_mask = np.in1d(gl_nnum, result.mesh.nnum, assume_unique=True)
             if not new_mask.any():  # pragma: no cover
-                raise RuntimeError('File %s not part of the distributed result'
-                                   % filenames[index])
+                raise RuntimeError(
+                    "File %s not part of the distributed result" % filenames[index]
+                )
             mask += new_mask
             self._results.append(result)
 
         if not np.all(mask):
-            filenames = '\n'.join([result.filename for result in self._results])
-            raise FileNotFoundError('Total nodes loaded from the individual result '
-                                    'files does not match the number in the global '
-                                    'index.  \n\nResult files found include:\n\n%s' %
-                                    filenames)
+            filenames = "\n".join([result.filename for result in self._results])
+            raise FileNotFoundError(
+                "Total nodes loaded from the individual result "
+                "files does not match the number in the global "
+                "index.  \n\nResult files found include:\n\n%s" % filenames
+            )
 
         # assemble the global mesh
         vtkappend = vtkAppendFilter()
@@ -98,10 +103,10 @@ class DistributedResult(Result):
         st = 0
         for i, result in enumerate(self._results):
             # map the merged node numbering with the individual mappings
-            result.grid['_dist_idx'] = np.arange(st, st + result.grid.n_points)
+            result.grid["_dist_idx"] = np.arange(st, st + result.grid.n_points)
             _result_idx = np.empty(result.grid.n_cells, np.int32)
             _result_idx[:] = i
-            result.grid.cell_arrays['_result_idx'] = _result_idx
+            result.grid.cell_data["_result_idx"] = _result_idx
             st += result.grid.n_points
             vtkappend.AddInputData(result.grid)
 
@@ -112,13 +117,13 @@ class DistributedResult(Result):
         # nodes are not sorted here
 
         # Map nodes from individual results to the global mesh
-        self._glb_idx = grid['_dist_idx']
+        self._glb_idx = grid["_dist_idx"]
 
         # start of each section of the grid
         # elem_split_ind = np.diff(self.grid['_result_idx']).nonzero()[0]
         # self._elem_split = np.hstack(([0], elem_split_ind))
 
-        elem = np.hstack(result.mesh._elem for result in self._results)
+        elem = np.hstack(list(result.mesh._elem for result in self._results))
         glb_elem_off = []
         for result in self._results:
             elem_off = result.mesh._elem_off
@@ -130,20 +135,21 @@ class DistributedResult(Result):
 
         # TODO: Add node and element components
         # resort node numbers
-        self._dist_sort_idx = np.argsort(grid['ansys_node_num'])
-        self._sorted_nnum = grid['ansys_node_num'][self._dist_sort_idx]
+        self._dist_sort_idx = np.argsort(grid["ansys_node_num"])
+        self._sorted_nnum = grid["ansys_node_num"][self._dist_sort_idx]
         nodes = grid.points[self._dist_sort_idx]
 
         glb_elem_off = np.hstack(glb_elem_off)
-        self._mesh = Mesh(self._sorted_nnum, nodes, elem, glb_elem_off,
-                          self._main_result.mesh.ekey)
-                          # node_comps=ncomp, elem_comps=ecomp)
+        self._mesh = Mesh(
+            self._sorted_nnum, nodes, elem, glb_elem_off, self._main_result.mesh.ekey
+        )
+        # node_comps=ncomp, elem_comps=ecomp)
 
         self.quadgrid = self._mesh._parse_vtk(fix_midside=False)
         self.grid = self.quadgrid.linear_copy()
 
         # self._neqv = self._resultheader['neqv']  # may not need this
-        self._eeqv = self.grid.cell_arrays['ansys_elem_num']
+        self._eeqv = self.grid.cell_data["ansys_elem_num"]
 
     @property
     def _main_result(self):
@@ -231,8 +237,9 @@ class DistributedResult(Result):
         """
         # check result exists
         if not self.available_results[result_type]:
-            raise ValueError('Result %s is not available in this result file'
-                             % result_type)
+            raise ValueError(
+                "Result %s is not available in this result file" % result_type
+            )
 
         # element header
         rnum = self.parse_step_substep(rnum)
@@ -249,80 +256,85 @@ class DistributedResult(Result):
 
         c = 0  # global cell index counter
         for result in self._results:
-            ele_ind_table, nodstr, etype, ptr_off = result._element_solution_header(rnum)
+            ele_ind_table, etype, ptr_off = result._element_solution_header(rnum)
             # we return c here since it is copied to C, not passed by reference
-            c = read_nodal_values_dist(result.filename,
-                                       self.grid.celltypes,
-                                       ele_ind_table,
-                                       offset,
-                                       cells,
-                                       nitem,
-                                       n_points,
-                                       nodstr,
-                                       etype,
-                                       self._mesh.etype,
-                                       result_index,
-                                       ptr_off,
-                                       ncount,
-                                       data,
-                                       c)
+            c = read_nodal_values_dist(
+                result.filename,
+                self.grid.celltypes,
+                ele_ind_table,
+                offset,
+                cells,
+                nitem,
+                n_points,
+                self._nodstr,
+                etype,
+                self._mesh.etype,
+                result_index,
+                ptr_off,
+                ncount,
+                data,
+                c,
+            )
 
-        if result_type == 'ENS' and nitem != 6:
+        if result_type == "ENS" and nitem != 6:
             data = data[:, :6]
 
         if not np.any(ncount):
-            raise ValueError('Result file contains no %s records for result %d' %
-                             (element_index_table_info[result_type.upper()], rnum))
+            raise ValueError(
+                "Result file contains no %s records for result %d"
+                % (element_index_table_info[result_type.upper()], rnum)
+            )
 
         # average across nodes
         data /= ncount.reshape(-1, 1)
-        return self.grid.point_arrays['ansys_node_num'], data
+        return self.grid.point_data["ansys_node_num"], data
 
     @wraps(Result.element_solution_data)
     def element_solution_data(self, *args, **kwargs):
         """Accumulate the element solution individual results from each result"""
-        sort = kwargs.get('sort', True)
+        sort = kwargs.get("sort", True)
         glb_element_data = []
         for result in self._results:
-            element_data = result.element_solution_data(*args, is_dist_rst=True,
-                                                        **kwargs)
+            element_data = result.element_solution_data(
+                *args, is_dist_rst=True, **kwargs
+            )
             glb_element_data.extend(element_data)
 
         # Assemble and sort (args[0] is rnum)
-        _, nodstr, etype, _ = self._element_solution_header(args[0])
+        _, etype, _ = self._element_solution_header(args[0])
 
         enum = self._eeqv
         enode = []
-        nnode = nodstr[etype]
+        nnode = self._nodstr[etype]
         if sort:
             sidx = np.argsort(enum)
             enum = enum[sidx]
             glb_element_data = [glb_element_data[i] for i in sidx]
 
             for i in sidx:
-                enode.append(self._mesh.elem[i][10:10+nnode[i]])
+                enode.append(self._mesh.elem[i][10 : 10 + nnode[i]])
 
         else:
             for i in range(enum.size):
-                enode.append(self._mesh.elem[i][10:10+nnode[i]])
+                enode.append(self._mesh.elem[i][10 : 10 + nnode[i]])
 
         return enum, glb_element_data, enode
 
     @wraps(Result.element_stress)
     def element_stress(self, *args, **kwargs):
         """Accumulate the element solution individual results from each result"""
-        sort = kwargs.get('sort', True)
+        sort = kwargs.get("sort", True)
         glb_stress = []
         for result in self._results:
             element_data = result.element_stress(*args, is_dist_rst=True, **kwargs)
             glb_stress.extend(element_data)
 
         # Assemble and sort (args[0] is rnum)
-        _, nodstr, etype, _ = self._element_solution_header(args[0])
+        _, etype, _ = self._element_solution_header(args[0])
 
         enum = self._eeqv
         enode = []
-        nnode = nodstr[etype]
+        nnode = self._nodstr[etype]
 
         if sort:
             sidx = np.argsort(enum)
@@ -330,16 +342,17 @@ class DistributedResult(Result):
             glb_element_data = [glb_stress[i] for i in sidx]
 
             for i in sidx:
-                enode.append(self._mesh.elem[i][10:10+nnode[i]])
+                enode.append(self._mesh.elem[i][10 : 10 + nnode[i]])
 
         else:
             for i in range(enum.size):
-                enode.append(self._mesh.elem[i][10:10+nnode[i]])
+                enode.append(self._mesh.elem[i][10 : 10 + nnode[i]])
 
         return enum, glb_element_data, enode
 
-    def plot_element_result(self, rnum, result_type, item_index,
-                            in_element_coord_sys=False, **kwargs):
+    def plot_element_result(
+        self, rnum, result_type, item_index, in_element_coord_sys=False, **kwargs
+    ):
         """Plot an element result.
 
         Parameters
@@ -399,20 +412,20 @@ class DistributedResult(Result):
         # check result exists
         result_type = result_type.upper()
         if not self.available_results[result_type]:
-            raise ValueError('Result %s is not available in this result file'
-                             % result_type)
+            raise ValueError(
+                "Result %s is not available in this result file" % result_type
+            )
 
         if result_type not in ELEMENT_INDEX_TABLE_KEYS:
             raise ValueError('Result "%s" is not an element result' % result_type)
 
         bsurfs = []
         for result in self._results:
-            bsurf = result._extract_surface_element_result(rnum,
-                                                           result_type,
-                                                           item_index,
-                                                           in_element_coord_sys)
+            bsurf = result._extract_surface_element_result(
+                rnum, result_type, item_index, in_element_coord_sys
+            )
             bsurfs.append(bsurf)
 
         desc = self.available_results.description[result_type].capitalize()
-        kwargs.setdefault('stitle', desc)
-        return pv.plot(bsurfs, scalars='_scalars', **kwargs)
+        kwargs.setdefault("stitle", desc)
+        return pv.plot(bsurfs, scalars="_scalars", **kwargs)
