@@ -1,30 +1,44 @@
 """Plotting helper for MAPDL using pyvista"""
-import pyvista as pv
 import numpy as np
+import pyvista as pv
 
 from ansys.mapdl.reader.misc import unique_rows
 
 
-def general_plotter(title, meshes, points, labels,
-                    cpos=None,
-                    show_bounds=False, show_axes=True,
-                    background=None, off_screen=None,
-                    screenshot=False,
-                    window_size=None,
-                    notebook=None,
-                    # add_mesh kwargs:
-                    color='w',
-                    show_edges=None, edge_color=None, point_size=5.0,
-                    line_width=None, opacity=1.0, flip_scalars=False,
-                    lighting=None, n_colors=256,
-                    interpolate_before_map=True, cmap=None,
-                    render_points_as_spheres=False, render_lines_as_tubes=False,
-                    stitle=None,
-                    smooth_shading=None,
-                    # labels kwargs
-                    font_size=None,
-                    font_family=None,
-                    text_color=None):
+def general_plotter(
+    title,
+    meshes,
+    points,
+    labels,
+    cpos=None,
+    show_bounds=False,
+    show_axes=True,
+    background=None,
+    off_screen=None,
+    screenshot=False,
+    window_size=None,
+    notebook=None,
+    # add_mesh kwargs:
+    color="w",
+    show_edges=None,
+    edge_color=None,
+    point_size=5.0,
+    line_width=None,
+    opacity=1.0,
+    flip_scalars=False,
+    lighting=None,
+    n_colors=256,
+    interpolate_before_map=True,
+    cmap=None,
+    render_points_as_spheres=False,
+    render_lines_as_tubes=False,
+    stitle=None,
+    smooth_shading=False,
+    # labels kwargs
+    font_size=None,
+    font_family=None,
+    text_color=None,
+):
     """General pyansys plotter for APDL geometry and meshes.
 
     Parameters
@@ -131,42 +145,59 @@ def general_plotter(title, meshes, points, labels,
         pl.set_background(background)
 
     for point in points:
-        pl.add_points(point['points'],
-                      scalars=point.get('scalars', None),
-                      color=color,
-                      show_edges=show_edges, edge_color=edge_color,
-                      point_size=point_size, line_width=line_width,
-                      opacity=opacity, flip_scalars=flip_scalars,
-                      lighting=lighting, n_colors=n_colors,
-                      interpolate_before_map=interpolate_before_map,
-                      cmap=cmap, render_points_as_spheres=render_points_as_spheres,
-                      render_lines_as_tubes=render_lines_as_tubes)
+        pl.add_points(
+            point["points"],
+            scalars=point.get("scalars", None),
+            color=color,
+            show_edges=show_edges,
+            edge_color=edge_color,
+            point_size=point_size,
+            line_width=line_width,
+            opacity=opacity,
+            flip_scalars=flip_scalars,
+            lighting=lighting,
+            n_colors=n_colors,
+            interpolate_before_map=interpolate_before_map,
+            cmap=cmap,
+            render_points_as_spheres=render_points_as_spheres,
+            render_lines_as_tubes=render_lines_as_tubes,
+        )
 
     for mesh in meshes:
-        pl.add_mesh(mesh['mesh'],
-                    scalars=mesh.get('scalars', None),
-                    stitle=mesh.get('stitle', None),
-                    color=mesh.get('color', color),
-                    show_edges=show_edges, edge_color=edge_color,
-                    smooth_shading=smooth_shading,
-                    point_size=point_size, line_width=line_width,
-                    opacity=opacity, flip_scalars=flip_scalars,
-                    lighting=lighting, n_colors=n_colors,
-                    interpolate_before_map=interpolate_before_map,
-                    cmap=cmap, render_points_as_spheres=render_points_as_spheres,
-                    render_lines_as_tubes=render_lines_as_tubes)
+        pl.add_mesh(
+            mesh["mesh"],
+            scalars=mesh.get("scalars", None),
+            stitle=mesh.get("stitle", None),
+            color=mesh.get("color", color),
+            show_edges=show_edges,
+            edge_color=edge_color,
+            smooth_shading=smooth_shading,
+            point_size=point_size,
+            line_width=line_width,
+            opacity=opacity,
+            flip_scalars=flip_scalars,
+            lighting=lighting,
+            n_colors=n_colors,
+            interpolate_before_map=interpolate_before_map,
+            cmap=cmap,
+            render_points_as_spheres=render_points_as_spheres,
+            render_lines_as_tubes=render_lines_as_tubes,
+        )
 
     for label in labels:
         # verify points are not duplicates
-        points, idx, _ = unique_rows(np.array(label['points']))
-        labels = np.array(label['labels'])[idx].tolist()
+        points, idx, _ = unique_rows(np.array(label["points"]))
+        labels = np.array(label["labels"])[idx].tolist()
 
-        pl.add_point_labels(points,
-                            labels,
-                            show_points=False, shadow=False,
-                            font_size=font_size,
-                            font_family=font_family,
-                            text_color=text_color)
+        pl.add_point_labels(
+            points,
+            labels,
+            show_points=False,
+            shadow=False,
+            font_size=font_size,
+            font_family=font_family,
+            text_color=text_color,
+        )
 
     if stitle is not None:
         pl.add_scalar_bar(title=stitle)
