@@ -46,76 +46,75 @@ nodal data
 
 
 DESCRIPTION = {
-    'EMS': 'miscellaneous summable items(normally includes face pressures)',
-    'ENF': 'nodal forces',
-    'ENS': 'nodal stresses',
-    'ENG': 'element energies and volume',
-    'EGR': 'nodal gradients',
-    'EEL': 'nodal elastic strains',
-    'EPL': 'nodal plastic strains',
-    'ECR': 'nodal creep strains',
-    'ETH': 'nodal thermal strains (includes swelling strains)',
-    'EUL': 'element euler angles',
-    'EFX': 'nodal fluxes',
-    'ELF': 'nodal forces generated by the element, e.g. lorentz or maxwell forces',
-    'EMN': 'miscellaneous nonsummable items',
-    'ECD': 'nodal current densities',
-    'ENL': 'nodal nonlinear items, e.g. equivalent plastic strains',
-    'EHC': 'element joule heating',
-    'EPT': 'nodal temperatures',
-    'SURF': 'face surface stresses',
-    'EDI': 'nodal diffusion strains',
-    'ESTR': 'POST1 element (ETABLE) data',
-    'ECT': 'nodal contact items, e.g. penetration',
-    'EXYZ': 'integration point locations',
-    'EBAC': 'back stresses(for kinematic hardening)',
-    'ESVR': 'saved variables from usermat',
+    "EMS": "miscellaneous summable items(normally includes face pressures)",
+    "ENF": "nodal forces",
+    "ENS": "nodal stresses",
+    "ENG": "element energies and volume",
+    "EGR": "nodal gradients",
+    "EEL": "nodal elastic strains",
+    "EPL": "nodal plastic strains",
+    "ECR": "nodal creep strains",
+    "ETH": "nodal thermal strains (includes swelling strains)",
+    "EUL": "element euler angles",
+    "EFX": "nodal fluxes",
+    "ELF": "nodal forces generated by the element, e.g. lorentz or maxwell forces",
+    "EMN": "miscellaneous nonsummable items",
+    "ECD": "nodal current densities",
+    "ENL": "nodal nonlinear items, e.g. equivalent plastic strains",
+    "EHC": "element joule heating",
+    "EPT": "nodal temperatures",
+    "SURF": "face surface stresses",
+    "EDI": "nodal diffusion strains",
+    "ESTR": "POST1 element (ETABLE) data",
+    "ECT": "nodal contact items, e.g. penetration",
+    "EXYZ": "integration point locations",
+    "EBAC": "back stresses(for kinematic hardening)",
+    "ESVR": "saved variables from usermat",
     # 'EMR': 'element material saved record',
-    'NSL': 'Nodal displacements',
-    'VSL': 'Nodal velocities',
-    'ASL': 'Nodal accelerations',
-    'RF': 'Nodal reaction forces',
+    "NSL": "Nodal displacements",
+    "VSL": "Nodal velocities",
+    "ASL": "Nodal accelerations",
+    "RF": "Nodal reaction forces",
 }
 
 
-class AvailableResults():
-
+class AvailableResults:
     def __init__(self, avail_bits, is_thermal):
         """Parse the available bits and determine if a given result is
         available.
         """
         self._avail_bits = avail_bits
         self._parsed_bits = {
-            'EMS': self.ems,
-            'ENF': self.enf,
-            'ENS': self.ens,
-            'ENG': self.eng,
-            'EGR': self.egr,
-            'EEL': self.eel,
-            'EPL': self.epl,
-            'ECR': self.ecr,
-            'ETH': self.eth,
-            'EUL': self.eul,
-            'EFX': self.efx,
-            'ELF': self.elf,
-            'EMN': self.emn,
-            'ECD': self.ecd,
-            'ENL': self.enl,
-            'EHC': self.ehc,
-            'EPT': self.ept,
-            'SURF': self.surf,
-            'EDI': self.edi,
-            'ESTR': self.estr,
-            'ECT': self.ect,
-            'EXYZ': self.exyz,
-            'EBAC': self.ebac,
-            'ESVR': self.esvr,
+            "EMS": self.ems,
+            "ENF": self.enf,
+            "ENS": self.ens,
+            "ENG": self.eng,
+            "EGR": self.egr,
+            "EEL": self.eel,
+            "EPL": self.epl,
+            "ECR": self.ecr,
+            "ETH": self.eth,
+            "EUL": self.eul,
+            "EFX": self.efx,
+            "ELF": self.elf,
+            "EMN": self.emn,
+            "ECD": self.ecd,
+            "ENL": self.enl,
+            "EHC": self.ehc,
+            "EPT": self.ept,
+            "SURF": self.surf,
+            "EDI": self.edi,
+            "ESTR": self.estr,
+            "ECT": self.ect,
+            "EXYZ": self.exyz,
+            "EBAC": self.ebac,
+            "ESVR": self.esvr,
             # 'EMR': self.emr,
-            'NSL': self.nsl,
-            'VSL': self.vsl,
-            'ASL': self.asl,
-            'RF': self.rf,
-            }
+            "NSL": self.nsl,
+            "VSL": self.vsl,
+            "ASL": self.asl,
+            "RF": self.rf,
+        }
 
         self._valid_results = []
         for key, value in self._parsed_bits.items():
@@ -124,7 +123,7 @@ class AvailableResults():
 
         self.description = DESCRIPTION
         if is_thermal:
-            self.description['NSL'] = 'Nodal temperatures'
+            self.description["NSL"] = "Nodal temperatures"
 
     def __getitem__(self, key):
         """Allow a key access"""
