@@ -31,7 +31,7 @@ print(rotor)
 # plane.
 #
 
-_ = rotor.plot_sectors(cpos='xy', stitle='Sector', smooth_shading=True, cmap='bwr')
+_ = rotor.plot_sectors(cpos="xy", stitle="Sector", smooth_shading=True, cmap="bwr")
 
 
 ###############################################################################
@@ -48,11 +48,11 @@ print(rotor.harmonic_indices)
 ###############################################################################
 # You can refer to the result set either using MAPDL's 1-based indexing
 # which is (Load-step, sub-step).
-_ = rotor.plot_nodal_displacement((2, 2), comp='norm', cpos='xy')
+_ = rotor.plot_nodal_displacement((2, 2), comp="norm", cpos="xy")
 
 ###############################################################################
 # Alternatively, you can refer to the results using cumulative indexing.
-_ = rotor.plot_nodal_displacement(10, comp='norm', cpos='xy')
+_ = rotor.plot_nodal_displacement(10, comp="norm", cpos="xy")
 
 
 ###############################################################################
@@ -61,7 +61,7 @@ _ = rotor.plot_nodal_displacement(10, comp='norm', cpos='xy')
 #
 # ANSYS has great documentation on harmonic indexing within their
 # internal documentation, but for the sake of completeness it will be repeated here.
-# 
+#
 # The harmonic index used within MAPDL is an integer that determines
 # the variation in the value of a single degree of freedom at points
 # spaced at a circumferential angle equal to the sector angle.
@@ -72,7 +72,7 @@ _ = rotor.plot_nodal_displacement(10, comp='norm', cpos='xy')
 # harmonic index of 5:
 
 rnum = 10
-print('Harmonic Index:', rotor.harmonic_indices[rnum])
+print("Harmonic Index:", rotor.harmonic_indices[rnum])
 
 
 # For the purpose of illustrating the nodal diameter content, we
@@ -83,8 +83,15 @@ print('Harmonic Index:', rotor.harmonic_indices[rnum])
 # as five, and only 5 lines can be drawn through the node lines
 # spanning the circumference of the rotor.
 r = 0.55
-_ = rotor.plot_nodal_displacement(rnum, comp='norm', cpos='xy', rng=[r, r + 1E-2],
-                                  stitle=None, add_text=False, cmap='bwr')
+_ = rotor.plot_nodal_displacement(
+    rnum,
+    comp="norm",
+    cpos="xy",
+    rng=[r, r + 1e-2],
+    stitle=None,
+    add_text=False,
+    cmap="bwr",
+)
 
 
 ###############################################################################
@@ -93,8 +100,15 @@ rnum = 3
 print(rotor.harmonic_indices[rnum])
 
 r = 0
-_ = rotor.plot_nodal_displacement(rnum, comp='z', cpos='xy', rng=[r, r + 1E-2],
-                                  add_text=False, stitle=None, cmap='bwr')
+_ = rotor.plot_nodal_displacement(
+    rnum,
+    comp="z",
+    cpos="xy",
+    rng=[r, r + 1e-2],
+    add_text=False,
+    stitle=None,
+    cmap="bwr",
+)
 
 
 # Therefore for this rotor with the number of sectors (N) equaling the
@@ -119,17 +133,17 @@ _ = rotor.plot_nodal_displacement(rnum, comp='z', cpos='xy', rng=[r, r + 1E-2],
 # - N is the number of sectors
 # - k is the harmonic index
 # - m is a set of integers from 0 to infinity
-# 
+#
 # Therefore, for a 6 sector rotor, the first 4 possible nodal diameters at
 # harmonic index 1 are:
 
 N = 6  # number of sectors
 k = 1  # harmonic index
-print('Nodal Diameters :')
+print("Nodal Diameters :")
 for m in range(5):
     if m > 0:
-        print('%4d' % (m*N - k))
-    print('%4d' % (m*N + k))
+        print("%4d" % (m * N - k))
+    print("%4d" % (m * N + k))
 
 ###############################################################################
 # Normally for a rotor where the number of blades per sector is 1,
@@ -146,17 +160,17 @@ n_blades = 24  # 4 blades per sector
 nodal_diameters = []
 for m in range(5):
     if m > 0:
-        nodal_diameters.append(m*N - k)
-    nodal_diameters.append(m*N + k)
+        nodal_diameters.append(m * N - k)
+    nodal_diameters.append(m * N + k)
 
-print('Nodal Diameters :')
+print("Nodal Diameters :")
 for d in nodal_diameters:
     if d > n_blades:  # ignore
         continue
-    elif d > n_blades//2:  # back alias
-        print('%4d' % -(n_blades-d))
+    elif d > n_blades // 2:  # back alias
+        print("%4d" % -(n_blades - d))
     else:
-        print('%4d' % d)
+        print("%4d" % d)
 
 
 ###############################################################################
@@ -168,8 +182,9 @@ for d in nodal_diameters:
 multi_blade_rotor = examples.download_academic_rotor_4blade_result()
 
 # plot all the sectors
-_ = multi_blade_rotor.plot_sectors(cpos='xy', stitle='Sector',
-                                   smooth_shading=True, cmap='bwr')
+_ = multi_blade_rotor.plot_sectors(
+    cpos="xy", stitle="Sector", smooth_shading=True, cmap="bwr"
+)
 
 ###############################################################################
 # We can then see that the harmonic indices for harmonic index 1
@@ -178,46 +193,42 @@ _ = multi_blade_rotor.plot_sectors(cpos='xy', stitle='Sector',
 print(multi_blade_rotor.harmonic_indices)
 
 hi_1 = np.nonzero(multi_blade_rotor.harmonic_indices == 1)[0]
-print('Cumulative indices matching harmonic index 1:', hi_1)
+print("Cumulative indices matching harmonic index 1:", hi_1)
 
 ###############################################################################
 # Cumulative result 8 and harmonic index 1
 rnum = 8
-text = 'Expected Nodal Diameter %2d' % nodal_diameters[0]
-_ = multi_blade_rotor.plot_nodal_displacement(rnum, comp='z',
-                                              cpos='xy', rng=[r, r + 1E-2],
-                                              add_text=text, stitle=None,
-                                              cmap='bwr')
+text = "Expected Nodal Diameter %2d" % nodal_diameters[0]
+_ = multi_blade_rotor.plot_nodal_displacement(
+    rnum, comp="z", cpos="xy", rng=[r, r + 1e-2], add_text=text, stitle=None, cmap="bwr"
+)
 
 
 ###############################################################################
 # Cumulative result 10 and harmonic index 5
 rnum = 10
-text = 'Expected Nodal Diameter %2d' % nodal_diameters[1]
-_ = multi_blade_rotor.plot_nodal_displacement(rnum, comp='z',
-                                              cpos='xy', rng=[r, r + 1E-2],
-                                              add_text=text, stitle=None,
-                                              cmap='bwr')
+text = "Expected Nodal Diameter %2d" % nodal_diameters[1]
+_ = multi_blade_rotor.plot_nodal_displacement(
+    rnum, comp="z", cpos="xy", rng=[r, r + 1e-2], add_text=text, stitle=None, cmap="bwr"
+)
 
 
 ###############################################################################
 # Cumulative result 13 and harmonic index 7
 rnum = 13
-text = 'Expected Nodal Diameter %2d' % nodal_diameters[2]
-_ = multi_blade_rotor.plot_nodal_displacement(rnum, comp='z',
-                                              cpos='xy', rng=[r, r + 1E-2],
-                                              add_text=text, stitle=None,
-                                              cmap='bwr')
+text = "Expected Nodal Diameter %2d" % nodal_diameters[2]
+_ = multi_blade_rotor.plot_nodal_displacement(
+    rnum, comp="z", cpos="xy", rng=[r, r + 1e-2], add_text=text, stitle=None, cmap="bwr"
+)
 
 
 ###############################################################################
 # Cumulative result 15 and harmonic index 11
 rnum = 15
-text = 'Expected Nodal Diameter %2d' % nodal_diameters[3]
-_ = multi_blade_rotor.plot_nodal_displacement(rnum, comp='z',
-                                              cpos='xy', rng=[r, r + 1E-2],
-                                              add_text=text, stitle=None,
-                                              cmap='bwr')
+text = "Expected Nodal Diameter %2d" % nodal_diameters[3]
+_ = multi_blade_rotor.plot_nodal_displacement(
+    rnum, comp="z", cpos="xy", rng=[r, r + 1e-2], add_text=text, stitle=None, cmap="bwr"
+)
 
 
 ###############################################################################
@@ -231,20 +242,32 @@ _ = multi_blade_rotor.plot_nodal_displacement(rnum, comp='z',
 # on the settings within the modal analysis.
 #
 # Set ``loop=True`` to allow for a continuous plot, and modify
-# ``nangles`` to allow for a smoother or faster plot.  See
+# ``n_frames`` to allow for a smoother or faster plot.  See
 # ``help(pyvista.plot)`` for additional keyword arguments.
 #
-_ = rotor.animate_nodal_displacement((3, 1), displacement_factor=0.03,
-                                     nangles=30, show_axes=False, background='w',
-                                     loop=False, add_text=False,
-                                     movie_filename='EO3_Mode1.gif')
+_ = rotor.animate_nodal_displacement(
+    (3, 1),
+    displacement_factor=0.03,
+    n_frames=30,
+    show_axes=False,
+    background="w",
+    loop=False,
+    add_text=False,
+    movie_filename="EO3_Mode1.gif",
+)
 
 
 ###############################################################################
 # Note how you can plot the backwards traveling wave by selecting the
 # second mode in the mode pair ``(3, 2)`` instead of ``(3, 1)``.
 #
-_ = rotor.animate_nodal_displacement((3, 2), displacement_factor=0.03,
-                                     nangles=30, show_axes=False, background='w',
-                                     loop=False, add_text=False,
-                                     movie_filename='EO3_Mode1.gif')
+_ = rotor.animate_nodal_displacement(
+    (3, 2),
+    displacement_factor=0.03,
+    n_frames=30,
+    show_axes=False,
+    background="w",
+    loop=False,
+    add_text=False,
+    movie_filename="EO3_Mode1.gif",
+)
