@@ -22,7 +22,7 @@ print(rotor)
 # the load step and sub step in fortran indexing.  You could have also
 # used the python cumulative index 3.
 # rotor._positive_cyclic_dir = True
-_ = rotor.plot_nodal_displacement((2, 2), 'x', cpos='xy')
+_ = rotor.plot_nodal_displacement((2, 2), "x", cpos="xy")
 
 
 ###############################################################################
@@ -48,7 +48,6 @@ _ = rotor.plot_nodal_displacement((2, 2), 'x', cpos='xy')
 nnum, strain = rotor.nodal_elastic_strain(3, full_rotor=True)
 
 
-
 ###############################################################################
 # Plot the nodal elastic strain in the "Z" direction for result ``(5, 2)``.
 #
@@ -64,8 +63,9 @@ nnum, strain = rotor.nodal_elastic_strain(3, full_rotor=True)
 # manually closing the plotting screen, set ``off_screen=True``.  This
 # can help you automate saving screenshots.
 
-_ = rotor.plot_nodal_elastic_strain((5, 2), 'Z', show_displacement=True,
-                                    displacement_factor=0.01)
+_ = rotor.plot_nodal_elastic_strain(
+    (5, 2), "Z", show_displacement=True, displacement_factor=0.01
+)
 
 
 ###############################################################################
@@ -75,8 +75,9 @@ _ = rotor.plot_nodal_elastic_strain((5, 2), 'Z', show_displacement=True,
 # result ``(5, 2)``
 #
 # Available stress components are ``['Y', 'Z', 'XY', 'YZ', 'XZ']``
-_ = rotor.plot_nodal_stress((5, 1), 'Z', show_displacement=True,
-                            displacement_factor=0.01)
+_ = rotor.plot_nodal_stress(
+    (5, 1), "Z", show_displacement=True, displacement_factor=0.01
+)
 
 
 ###############################################################################
@@ -84,5 +85,6 @@ _ = rotor.plot_nodal_stress((5, 1), 'Z', show_displacement=True,
 # shows the principal stress for result ``(5, 2)``.
 #
 # Available stress components are ``['S1', 'S2', 'S3', 'SINT', 'SEQV']``.
-_ = rotor.plot_principal_nodal_stress((5, 2), 'SEQV', show_displacement=True,
-                                      displacement_factor=0.01)
+_ = rotor.plot_principal_nodal_stress(
+    (5, 2), "SEQV", show_displacement=True, displacement_factor=0.01
+)
