@@ -1,7 +1,7 @@
 import os
 
-import pytest
 import numpy as np
+import pytest
 
 from ansys.mapdl import reader as pymapdl_reader
 from ansys.mapdl.reader import examples
@@ -9,12 +9,12 @@ from ansys.mapdl.reader import examples
 test_path = os.path.dirname(os.path.abspath(__file__))
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture(scope="module")
 def result():
     return pymapdl_reader.read_binary(examples.rstfile)
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture(scope="module")
 def archive():
     return pymapdl_reader.Archive(examples.hexarchivefile)
 
@@ -33,7 +33,7 @@ def test_geometry_nodenum(result, archive):
 
 
 def test_results_displacement(result):
-    textfile = os.path.join(test_path, 'prnsol_u.txt')
+    textfile = os.path.join(test_path, "prnsol_u.txt")
     nnum, r_values = result.nodal_solution(0)
     a_values = np.loadtxt(textfile, skiprows=2)[:, 1:4]
     assert np.allclose(r_values, a_values)
@@ -41,17 +41,17 @@ def test_results_displacement(result):
 
 def test_results_stress(result):
     _, r_values = result.nodal_stress(0)
-    textfile = os.path.join(test_path, 'prnsol_s.txt')
+    textfile = os.path.join(test_path, "prnsol_s.txt")
     a_values = np.loadtxt(textfile, skiprows=2)[:, 1:]
 
     # ignore nan
     nanmask = ~np.isnan(r_values).any(1)
-    assert np.allclose(r_values[nanmask], a_values, atol=1E-1)
+    assert np.allclose(r_values[nanmask], a_values, atol=1e-1)
 
 
 def test_results_pstress(result):
     r_nnum, r_values = result.principal_nodal_stress(0)
-    textfile = os.path.join(test_path, 'prnsol_s_prin.txt')
+    textfile = os.path.join(test_path, "prnsol_s_prin.txt")
     a_values = np.loadtxt(textfile, skiprows=2)[:, 1:]
 
     # ignore nan
