@@ -1,9 +1,10 @@
 import os
+
 from ansys.mapdl import reader as pymapdl_reader
 
 TEST_PATH = os.path.dirname(os.path.abspath(__file__))
-TESTFILES_PATH = os.path.join(TEST_PATH, 'testfiles')
-BEAM44_RST = os.path.join(TESTFILES_PATH, 'beam44.rst')
+TESTFILES_PATH = os.path.join(TEST_PATH, "testfiles")
+BEAM44_RST = os.path.join(TESTFILES_PATH, "beam44.rst")
 
 
 def test_beam44():
